10.GeneSetTests {limma}	R Documentation

Topic: Gene Set Tests

Description

This page gives an overview of the LIMMA functions for gene set testing and pathway analysis.

roast

Self-contained gene set testing for one set. Uses zscoreT to normalize t-statistics.

mroast

Self-contained gene set testing for many sets. Uses zscoreT to normalize t-statistics.

fry

Fast approximation to mroast, especially useful when heteroscedasticity of genes can be ignored.

camera

Competitive gene set testing.

romer and topRomer

Gene set enrichment analysis.

ids2indices

Convert gene sets consisting of vectors of gene identifiers into a list of indices suitable for use in the above functions.

alias2Symbol, alias2SymbolTable and alias2SymbolUsingNCBI

Convert gene symbols or aliases to current official symbols.

geneSetTest or wilcoxGST

Simple gene set testing based on gene or probe permutation.

barcodeplot

Enrichment plot of a gene set.

goana and topGO

Gene ontology over-representation analysis of gene lists using Entrez Gene IDs. goana can work directly on a fitted model object or on one or more lists of genes.

kegga and topKEGG

KEGG pathway over-representation analysis of gene lists using Entrez Gene IDs. kegga can work directly on a fitted model object or on one or more lists of genes.

Author(s)

Gordon Smyth

See Also

01.Introduction, 02.Classes, 03.ReadingData, 04.Background, 05.Normalization, 06.LinearModels, 07.SingleChannel, 08.Tests, 09.Diagnostics, 10.GeneSetTests, 11.RNAseq

[Package limma version 3.40.6 Index]