10.Other {limma}	R Documentation

Other Functions

Description

This page describes some functions not covered in the previous numbered pages, so far only blockDiag and poolVar which are not used in the package yet but are part of the development of methods to handle technical and biological replicates.

Author(s)

Gordon Smyth

[Package limma version 2.14.7 Index]