11.RNAseq {limma}	R Documentation

Topic: Analysis of RNA-seq Data

Description

This page gives an overview of LIMMA functions to analyze RNA-seq data.

voom

Transform RNA-seq or ChIP-seq counts to log counts per million (log-cpm) with associated precision weights. After this tranformation, RNA-seq or ChIP-seq data can be analyzed using the same functions as would be used for microarray data.

voomWithQualityWeights

Combines the functionality of voom and arrayWeights.

diffSplice

Test for differential exon usage between experimental conditions.

topSplice

Show a data.frame of top results from diffSplice.

plotSplice

Plot results from diffSplice.

plotExons

Plot logFC for individual exons for a given gene.

References

Law, CW, Chen, Y, Shi, W, Smyth, GK (2014). Voom: precision weights unlock linear model analysis tools for RNA-seq read counts. Genome Biology 15, R29. http://genomebiology.com/2014/15/2/R29

Ritchie, ME, Phipson, B, Wu, D, Hu, Y, Law, CW, Shi, W, and Smyth, GK (2015). limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic Acids Research 43, e47. http://nar.oxfordjournals.org/content/43/7/e47

See Also

See also the edgeR package for normalization and data summaries of RNA-seq data, as well as for alternative differential expression methods based on the negative binomial distribution. voom accepts DGEList objects and normalization factors from edgeR.

01.Introduction, 02.Classes, 03.ReadingData, 04.Background, 05.Normalization, 06.LinearModels, 07.SingleChannel, 08.Tests, 09.Diagnostics, 10.GeneSetTests, 11.RNAseq

[Package limma version 3.40.6 Index]