PrintLayout {limma}	R Documentation

Print Layout - class

Description

A list-based class for storing information about the process used to print spots on a microarray.

PrintLayout objects can be created using getLayout. The printer component of an RGList or MAList object is of this class.

Slots/List Components

Objects of this class contains no slots but should contain the following list components:

ngrid.r:	 number of grid rows on the arrays

ngrid.c:	 number of grid columns on the arrays

nspot.r:	 number of rows of spots in each grid

nspot.c:	 number of columns of spots in each grid

ndups:	 number of duplicates of each DNA clone, i.e., number of times print-head dips into each well of DNA

spacing:	 number of spots between duplicate spots. Only applicable if ndups>1. spacing=1 for side-by-side spots by rows, spacing=nspot.c for side-by-side spots by columns, spacing=ngrid.r*ngrid.c*nspot.r*nspot.c/2 for duplicate spots in top and bottom halves of each array.

npins:	 actual number of pins or tips on the print-head

start:	 character string giving position of the spot printed first in each grid. Choices are "topleft" or "topright" and partial matches are accepted.

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of all the classes defined by this package.

Examples


#  Settings for Swirl and ApoAI example data sets in User's Guide

printer <- list(ngrid.r=4, ngrid.c=4, nspot.r=22, nspot.c=24,
                ndups=1, spacing=1, npins=16, start="topleft")

#  Typical settings at the Australian Genome Research Facility

#  Full pin set, duplicates side-by-side on same row
printer <- list(ngrid.r=12, ngrid.c=4, nspot.r=20, nspot.c=20,
                ndups=2, spacing=1, npins=48, start="topright")

#  Half pin set, duplicates in top and lower half of slide
printer <- list(ngrid.r=12, ngrid.c=4, nspot.r=20, nspot.c=20,
                ndups=2, spacing=9600, npins=24, start="topright")

[Package limma version 3.40.6 Index]