blockDiag {limma}	R Documentation

Block Diagonal Matrix

Description

Form a block diagonal matrix from the given blocks.

Usage


blockDiag(...)

Arguments

... 	 numeric matrices

Value

A block diagonal matrix with dimensions equal to the sum of the input dimensions

Author(s)

Gordon Smyth

See Also

10.Other

Examples


a <- matrix(1,3,2)
b <- matrix(2,2,2)
blockDiag(a,b)

[Package limma version 2.14.7 Index]