bwss.matrix {limma}	R Documentation

Between and within sums of squares for matrix

Description

Sums of squares between and within the columns of a matrix. Allows for missing values. This function is called by the anova method for MAList objects.

Usage

bwss.matrix(x)

Arguments

x
a numeric matrix.

Details

This is equivalent to a one-way analysis of variance where the columns of the matrix are the groups. If x is a matrix then bwss.matrix(x) is the same as bwss(x,col(x)) except for speed of execution.

Value

A list with components

bss
sums of squares between the column means.

wss
sums of squares within the column means.

bdf
degrees of freedom corresponding to bss.

wdf
degrees of freedom corresponding to wss.

Author(s)

Gordon Smyth

See Also

bwss, anova.MAList

[Package limma version 3.40.6 Index]