changeLog {limma}	R Documentation

Limma Change Log

Description

Write as text the most recent changes from the limma package changelog.

Usage


changeLog(n=20)

Arguments

n
integer, number of lines to write of changelog.

Value

No value is produced, but a number of lines of text are written to standard output.

Author(s)

Gordon Smyth

See Also

01.Introduction

Examples


changeLog()

[Package limma version 3.40.6 Index]