diffSplice {limma}	R Documentation

Test for Differential Splicing

Description

Given a linear model fit at the exon level, test for differences in exon retention between experimental conditions.

Usage


diffSplice(fit, geneid, exonid=NULL, robust=FALSE, verbose=TRUE)

Arguments

fit
an MArrayLM fitted model object produced by lmFit or contrasts.fit. Rows should correspond to exons.

geneid
gene identifiers. Either a vector of length nrow(fit) or the name of the column of fit$genes containing the gene identifiers. Rows with the same ID are assumed to belong to the same gene.

exonid
exon identifiers. Either a vector of length nrow(fit) or the name of the column of fit$genes containing the exon identifiers.

robust
logical, should the estimation of the empirical Bayes prior parameters be robustified against outlier sample variances?

verbose
logical, if TRUE some diagnostic information about the number of genes and exons is output.

Details

This function tests for differential exon usage for each gene and for each column of fit.

Testing for differential exon usage is equivalent to testing whether the log-fold-changes in the fit differ between exons for the same gene. Two different tests are provided. The first is an F-test for differences between the log-fold-changes. The other is a series of t-tests in which each exon is compared to the average of all other exons for the same gene. The exon-level t-tests are converted into a genewise test by adjusting the p-values for the same gene by Simes method. The minimum adjusted p-value is then used for each gene.

This function can be used on data from an exon microarray or can be used in conjunction with voom for exon-level RNA-seq counts.

Value

An object of class MArrayLM containing both exon level and gene level tests. Results are sorted by geneid and by exonid within gene.

coefficients
numeric matrix of coefficients of same dimensions as fit. Each coefficient is the difference between the log-fold-change for that exon versus the average log-fold-change for all other exons for the same gene.

t
numeric matrix of moderated t-statistics, of same dimensions as fit.

p.value
numeric vector of p-values corresponding to the t-statistics

genes
data.frame of exon annotation

genecolname
character string giving the name of the column of genes containing gene IDs

gene.F
numeric matrix of moderated F-statistics, one row for each gene.

gene.F.p.value
numeric matrix of p-values corresponding to gene.F

gene.simes.p.value
numeric matrix of Simes adjusted p-values, one row for each gene.

gene.genes
data.frame of gene annotation.

Author(s)

Gordon Smyth and Charity Law

See Also

topSplice, plotSplice

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples


## Not run: 
v <- voom(dge,design)
fit <- lmFit(v,design)
ex <- diffSplice(fit,geneid="EntrezID")
topSplice(ex)
plotSplice(ex)

## End(Not run)

[Package limma version 3.40.6 Index]