dimnames {limma}	R Documentation

Retrieve the Dimension Names of an RGList, MAList, EList, EListRaw or MArrayLM Object

Description

Retrieve the dimension names of a microarray data object.

Usage


## S3 method for class 'RGList'
dimnames(x)
## S3 replacement method for class 'RGList'
dimnames(x) <- value

Arguments

x
an object of class RGList, MAList, EList, EListRaw or (not for assignment) MArrayLM

value
a possible value for dimnames(x): see dimnames

Details

The dimension names of a microarray object are the same as those of the most important matrix component of that object.

A consequence is that rownames and colnames will work as expected.

Value

Either NULL or a list of length 2. If a list, its components are either NULL or a character vector the length of the appropriate dimension of x.

Author(s)

Gordon Smyth

See Also

dimnames in the base package.

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 3.40.6 Index]