dnormexp {limma}	R Documentation

The Normal + Exponential Convolution Distribution

Description

Density of the normal + exponential convolution distribution. This function is called by normexp.fit.

Usage


dnormexp(x,normmean,normsd,expmean,log=FALSE)
dnormexp.saddle(x,normmean,normsd,expmean,log=FALSE,secondorder=TRUE)

Arguments

x 	 numeric vector of quantiles

normmean 	 mean of normal distribution

normsd 	 standard deviation of normal distribution

expmean 	 mean of exponential distribution

log 	 logical, if TRUE, density values d are given as log(d)

secondorder 	 logical, if TRUE the second order saddle-point approximation is used

Details

dnormexp computes the exact density but may fail for some extreme parameter values. dnormexp.saddle computes the saddle-point approximation to the density.

Value

numeric vector of density or log-density values.

Author(s)

Jeremy Silver and Gordon Smyth

See Also

An overview of background correction functions is given in 04.Background.

[Package limma version 2.14.7 Index]