ebayes {limma}	R Documentation

Empirical Bayes Statistics for Differential Expression (Deprecated)

Description

Given a microarray linear model fit, compute moderated t-statistics, moderated F-statistic, and log-odds of differential expression by empirical Bayes moderation of the standard errors towards a common value. This function is now deprecated. Please use eBayes instead.

Usage

ebayes(fit, proportion=0.01, stdev.coef.lim=c(0.1,4),
       trend=FALSE, robust=FALSE, winsor.tail.p=c(0.05,0.1))

Arguments

fit
an MArrayLM fitted model object produced by lmFit or contrasts.fit. For ebayes only, fit can alternatively be an unclassed list produced by lm.series, gls.series or mrlm containing components coefficients, stdev.unscaled, sigma and df.residual.

proportion
numeric value between 0 and 1, assumed proportion of genes which are differentially expressed

stdev.coef.lim
numeric vector of length 2, assumed lower and upper limits for the standard deviation of log2-fold-changes for differentially expressed genes

trend
logical, should an intensity-trend be allowed for the prior variance? Default is that the prior variance is constant.

robust
logical, should the estimation of df.prior and var.prior be robustified against outlier sample variances?

winsor.tail.p
numeric vector of length 1 or 2, giving left and right tail proportions of x to Winsorize. Used only when robust=TRUE.

Details

This function is now deprecated. Please use eBayes instead.

Value

A list with components:

t
numeric vector or matrix of moderated t-statistics

p.value
numeric vector of p-values corresponding to the t-statistics

s2.prior
estimated prior value for sigma^2. A vector if covariate is non-NULL, otherwise a scalar.

df.prior
degrees of freedom associated with s2.prior

df.total
numeric vector of total degrees of freedom associated with t-statistics and p-values. Equal to df.prior+df.residual or sum(df.residual), whichever is smaller.

s2.post
numeric vector giving the posterior values for sigma^2

lods
numeric vector or matrix giving the log-odds of differential expression (natural log scale).

var.prior
estimated prior value for the variance of the log2-fold-change for differentially expressed gene

Author(s)

Gordon Smyth

References

Phipson, B, Lee, S, Majewski, IJ, Alexander, WS, and Smyth, GK (2016). Robust hyperparameter estimation protects against hypervariable genes and improves power to detect differential expression. Annals of Applied Statistics 10, 946-963. http://projecteuclid.org/euclid.aoas/1469199900

Smyth, G. K. (2004). Linear models and empirical Bayes methods for assessing differential expression in microarray experiments. Statistical Applications in Genetics and Molecular Biology 3, Article 3. http://www.statsci.org/smyth/pubs/ebayes.pdf

See Also

eBayes

[Package limma version 3.38.2 Index]