exprs.MA {limma}	R Documentation

Extract Log-Expression Matrix from MAList

Description

Extract the matrix of log-expression values from an MAList object.

Usage


exprs.MA(MA)

Arguments

MA
an MAList object.

Details

Converts M and A-values to log-expression values. The output matrix will have two columns for each array, in the order green, red for each array.

This contrasts with as.matrix.MAList which extracts the M-values only, or RG.MA which converts to expression values in RGList form.

Value

A numeric matrix with twice the columns of the input.

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 3.40.6 Index]