fitGammaIntercept {limma}	R Documentation

Fit Intercept to Vector of Gamma Distributed Variates

Description

Fit Intercept to Vector of Gamma Distributed Variates

Usage


fitGammaIntercept(y,offset=0,maxit=1000)

Arguments

y
numeric vector of positive response values.

offset
numeric vector giving known part of the expected value of y. Can be a single value, or else a vector of the same length as y.

maxit
maximum number of Newton iterations to be done.

Details

The values y are assumed to follow a gamma distribution with common shape parameter and with expected values given by x+offset. The function implements a globally convergent Newton iteration to estimate x.

Value

Numeric value giving intercept.

Author(s)

Gordon Smyth and Belinda Phipson

References

Phipson, B. (2013). Empirical Bayes modelling of expression profiles and their associations. PhD Thesis. University of Melbourne, Australia.

See Also

This function is called by genas.

Examples


offset <- runif(10)
x <- 9
mu <- x+offset
y <- rgamma(10,shape=20,scale=mu/20)
fitGammaIntercept(y,offset=offset)

[Package limma version 3.40.6 Index]