ids2indices {limma}	R Documentation

Convert Gene Identifiers to Indices for Gene Sets

Description

Make a list of gene identifiers into a list of indices for gene sets.

Usage


ids2indices(gene.sets, identifiers, remove.empty=TRUE)

Arguments

gene.sets
list of character vectors, each vector containing the gene identifiers for a set of genes.

identifiers
character vector of gene identifiers.

remove.empty
logical, should sets of size zero be removed from the output?

Details

This function used to create input for romer, mroast and camera function. Typically, identifiers is the vector of Entrez Gene IDs, and gene.sets is obtained constructed from a database of gene sets, for example a representation of the Molecular Signatures Database (MSigDB) downloaded from http://bioinf.wehi.edu.au/software/MSigDB.

Value

list of integer vectors, each vector containing the indices of a gene set in the vector identifiers.

Author(s)

Gordon Smyth and Yifang Hu

See Also

romer, mroast, camera

There is a topic page on 10.GeneSetTests.

Examples



## Not run: 

download.file("http://bioinf.wehi.edu.au/software/MSigDB/human_c2_v5p2.rdata", 
	      "human_c2_v5p2.rdata", mode = "wb")

load("human_c2_v5p2.rdata")
c2.indices <- ids2indices(Hs.c2, y$genes$GeneID)
camera(y, c2.indices, design)


## End(Not run)


[Package limma version 3.40.6 Index]