imageplot3by2 {limma}	R Documentation

Write Imageplots to Files

Description

Write imageplots to files in PNG format, six plots to a file in a 3 by 2 grid arrangement.

Usage


imageplot3by2(RG, z="Gb", prefix=paste("image",z,sep="-"), path=NULL,
              zlim=NULL, common.lim=TRUE, ...)

Arguments

RG
an RGList or MAList object, or any list with component named by z

z
character string giving name of component of RG to plot

prefix
character string giving prefix to attach to file names

path
character string specifying directory for output files

zlim
numeric vector of length 2, giving limits of response vector to be associated with saturated colors

common.lim
logical, should all plots on a page use the same axis limits

...
any other arguments are passed to imageplot

Details

At the time of writing, this function writes plots in PNG format in an arrangement optimized for A4-sized paper.

Value

No value is returned, but one or more files are written to the working directory. The number of files is determined by the number of columns of RG.

Author(s)

Gordon Smyth

See Also

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 3.40.6 Index]