limmaUsersGuide {limma}	R Documentation

View Limma User's Guide

Description

Finds the location of the Limma User's Guide and optionally opens it.

Usage


limmaUsersGuide(view=TRUE)

Arguments

view
logical, should the document be opened using the default PDF document reader?

Details

The function vignette("limma") will find the short limma Vignette which describes how to obtain the Limma User's Guide. The User's Guide is not itself a true vignette because it is not automatically generated using Sweave during the package build process. This means that it cannot be found using vignette, hence the need for this special function.

If the operating system is other than Windows, then the PDF viewer used is that given by Sys.getenv("R_PDFVIEWER"). The PDF viewer can be changed using Sys.putenv(R_PDFVIEWER=).

This function is used by drop-down Vignettes menu when the Rgui interface for Windows is used.

Value

Character string giving the file location.

Author(s)

Gordon Smyth

See Also

vignette, openPDF, openVignette, Sys.getenv, Sys.putenv

Examples


limmaUsersGuide(view=FALSE)

[Package limma version 3.40.6 Index]