logcosh {limma}	R Documentation

Logarithm of cosh

Description

Compute log(cosh(x)) without floating overflow or underflow

Usage


logcosh(x)

Arguments

x
a numeric vector or matrix.

Details

The computation uses asymptotic expressions for very large or very small arguments. For intermediate arguments, log(cosh(x)) is returned.

Value

Numeric vector or matrix of same dimensions as x.

Author(s)

Gordon K Smyth

See Also

logsumexp

Examples


x <- c(1e-8,1e-7,1e-6,1e-5,1e-4,1,3,50,800)
logcosh(x)
log(cosh(x))

[Package limma version 3.40.6 Index]