MArrayLM-class {limma}	R Documentation

Microarray Linear Model Fit - class

Description

A list-based S4 class for storing the results of fitting gene-wise linear models to a set of microarrays. Objects are normally created by lmFit, and additional components are added by eBayes.

Components

MArrayLM objects do not contain any slots (apart from .Data) but they should contain the following list components:

coefficients 	 matrix containing fitted coefficients or contrasts

stdev.unscaled 	 matrix containing unscaled standard deviations of the coefficients or contrasts

sigma 	 numeric vector containing residual standard deviations for each gene

df.residual 	 numeric vector containing residual degrees of freedom for each gene

The following additional components may be created by lmFit:

Amean 	 numeric vector containing the average log-intensity for each probe over all the arrays in the original linear model fit. Note this vector does not change when a contrast is applied to the fit using contrasts.fit.

genes 	 data.frame containing probe annotation.

design 	 design matrix.

cov.coefficients 	 numeric matrix giving the unscaled covariance matrix of the estimable coefficients

pivot 	 integer vector giving the order of coefficients in cov.coefficients. Is computed by the QR-decomposition of the design matrix.

qr 	 QR-decomposition of the design matrix (if the fit involved no weights or missing values).

... 	 other components returned by lm.fit (if the fit involved no weights or missing values).

The following component may be added by contrasts.fit:

contrasts 	 numeric matrix defining contrasts of coefficients for which results are desired.

The following components may be added by eBayes:

s2.prior 	 numeric value giving empirical Bayes estimated prior value for residual variances

df.prior 	 numeric vector giving empirical Bayes estimated degrees of freedom associated with s2.prior for each gene

df.total 	 numeric vector giving total degrees of freedom used for each gene, usually equal to df.prior + df.residual.

s2.post 	 numeric vector giving posterior residual variances

var.prior 	 numeric vector giving empirical Bayes estimated prior variance for each true coefficient

F 	 numeric vector giving moderated F-statistics for testing all contrasts equal to zero

F.p.value 	 numeric vector giving p-value corresponding to F.stat

t 	 numeric matrix containing empirical Bayes t-statistics

Methods

MArrayLM objects will return dimensions and hence functions such as dim, nrow and ncol are defined. MArrayLM objects inherit a show method from the virtual class LargeDataObject.

The functions eBayes, decideTests and classifyTestsF accept MArrayLM objects as arguments.

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of all the classes defined by this package.

[Package limma version 3.40.6 Index]