normalizeForPrintorder {limma}	R Documentation

Print-Order Normalization

Description

Normalize intensity values on one or more spotted microarrays to adjust for print-order effects.

Usage


normalizeForPrintorder(object, layout, start="topleft", method = "loess",
                       separate.channels = FALSE, span = 0.1, plate.size = 32)
normalizeForPrintorder.rg(R, G, printorder, method = "loess", separate.channels = FALSE,
                          span = 0.1, plate.size = 32, plot = FALSE)
plotPrintorder(object, layout, start="topleft", slide = 1, method = "loess",
               separate.channels = FALSE, span = 0.1, plate.size = 32)

Arguments

object
an RGList or list object containing components R and G which are matrices containing the red and green channel intensities for a series of arrays

R
numeric vector containing red channel intensities for a single microarray

G
numeric vector containing the green channel intensities for a single microarray

layout
list specifying the printer layout, see PrintLayout-class

start
character string specifying where printing starts in each pin group. Choices are "topleft" or "topright".

printorder
numeric vector specifying order in which spots are printed. Can be computed from printorder(layout,start=start).

slide
positive integer giving the column number of the array for which a plot is required

method
character string, "loess" if a smooth loess curve should be fitted through the print-order trend or "plate" if plate effects are to be estimated

separate.channels
logical, TRUE if normalization should be done separately for the red and green channel and FALSE if the normalization should be proportional for the two channels

span
numerical constant between 0 and 1 giving the smoothing span for the loess the curve. Ignored if method="plate".

plate.size
positive integer giving the number of consecutive spots corresponding to one plate or plate pack. Ignored if method="loess".

plot
logical. If TRUE then a scatter plot of the print order effect is sent to the current graphics device.

Details

Print-order is associated with the 384-well plates used in the printing of spotted microarrays. There may be variations in DNA concentration or quality between the different plates. The may be variations in ambient conditions during the time the array is printed.

This function is intended to pre-process the intensities before other normalization methods are applied to adjust for variations in DNA quality or concentration and other print-order effects.

Printorder means the order in which spots are printed on a microarray. Spotted arrays are printed using a print head with an array of print-tips. Spots in the various tip-groups are printed in parallel. Printing is assumed to start in the top left hand corner of each tip-groups and to proceed right and down by rows, or else to start in the top right hand and to proceed left and down by rows. See printorder for more details. (WARNING: this is not always the case.) This is true for microarrays printed at the Australian Genome Research Facility but might not be true for arrays from other sources.

If object is an RGList then printorder is performed for each intensity in each array.

plotPrintorder is a non-generic function which calls normalizeForPrintorder with plot=TRUE.

Value

normalizeForPrintorder produces an RGList containing normalized intensities.

The function plotPrintorder or normalizeForPrintorder.rg with plot=TRUE returns no value but produces a plot as a side-effect.

normalizeForPrintorder.rg with plot=FALSE returns a list with the following components:

R
numeric vector containing the normalized red channel intensities

G
numeric vector containing the normalized red channel intensites

R.trend
numeric vector containing the fitted printorder trend for the red channel

G.trend
numeric vector containing the fitted printorder trend for the green channe

Author(s)

Gordon Smyth

References

Smyth, G. K. Print-order normalization of cDNA microarrays. March 2002. http://www.statsci.org/smyth/pubs/porder/porder.html

See Also

printorder.

An overview of LIMMA functions for normalization is given in 05.Normalization.

Examples


## Not run: 
plotPrintorder(RG,layout,slide=1,separate=TRUE)
RG <- normalizeForPrintorder(mouse.data,mouse.setup)

## End(Not run)

[Package limma version 3.40.6 Index]