normexp {limma}	R Documentation

Normal + Exponential Log-Likelihood

Description

Marginal log-likelihood of foreground values for the normal + exponential convolution model and its derivatives. These functions are called by normexp.fit and are not normally called directly by the user.

Usage


normexp.m2loglik(par,x)
normexp.m2loglik.saddle(par,x)
normexp.grad(par,x)

Arguments

par 	 numeric vector of parameters

x 	 numeric vector of (background corrected) intensities

Details

The parameter vector par holds the normal mean, the normal log-standard deviation and the exponential mean.

normexp.m2loglik computes minus twice the log-likelihood, and normexp.grad it is derivative, based on the $normal(μ,σ^2)+exponential(α)$ convolution model for the intensities. The elements of par are $μ$, $log(σ)$ and $log(α)$.

normexp.m2loglik is the saddle-point approximation to the log-logelihood, which is generally prefered because it is numerically more stable.

Value

normexp.m2loglik returns a numeric scalar holding minus-twice the log-likelihood. normexp.grad returns a numeric vector holding the derivatives with respect to the elements of par.

Author(s)

Jeremy Silver and Gordon Smyth

References

Ritchie, M. E., Silver, J., Oshlack, A., Silver, J., Holmes, M., Diyagama, D., Holloway, A., and Smyth, G. K. (2007). A comparison of background correction methods for two-colour microarrays. Bioinformatics http://bioinformatics.oxfordjournals.org/cgi/content/abstract/btm412

See Also

An overview of background correction functions is given in 04.Background.

[Package limma version 2.14.7 Index]