normexp.signal {limma}	R Documentation

Expected Signal Given Observed Foreground Under Normal+Exp Model

Description

Adjust foreground intensities for observed background using Normal+Exp Model. This function is called by backgroundCorrect and is not normally called directly by the user.

Usage


normexp.signal(par, x)

Arguments

par
numeric vector containing the parameters of the Normal+Exp distribution, see normexp.fit for details.

x
numeric vector of (background corrected) intensities

Details

In general the vector normmean is computed conditional on background at each spot.

Value

Numeric vector containing adjusted intensities.

Author(s)

Gordon Smyth

References

Ritchie, M. E., Silver, J., Oshlack, A., Silver, J., Holmes, M., Diyagama, D., Holloway, A., and Smyth, G. K. (2007). A comparison of background correction methods for two-colour microarrays. Bioinformatics 23, 2700-2707. http://bioinformatics.oxfordjournals.org/content/23/20/2700

Silver, JD, Ritchie, ME, and Smyth, GK (2009). Microarray background correction: maximum likelihood estimation for the normal-exponential convolution. Biostatistics 10, 352-363. http://biostatistics.oxfordjournals.org/content/10/2/352

See Also

normexp.fit

An overview of background correction functions is given in 04.Background.

Examples

# See normexp.fit

[Package limma version 3.40.6 Index]