plotSplice {limma}	R Documentation

Differential splicing plot

Description

Plot relative log-fold changes by exons for the specified gene and highlight the significantly spliced exons.

Usage


plotSplice(fit, coef=ncol(fit), geneid=NULL, genecolname=NULL, rank=1L, FDR = 0.05)

Arguments

fit
MArrayLM fit object produced by diffSplice.

coef
the coefficient (column) of fit for which differentially splicing is assessed.

geneid
character string, ID of the gene to plot.

genecolname
column name of fit$genes containing gene IDs. Defaults to fit$genecolname.

rank
integer, if geneid=NULL then this ranked gene will be plotted.

FDR
numeric, highlight exons as red dots with false discovery rate less than this cutoff. The FDR of the individual exon is calculated based on the exon-level t-statistics test for differences between each exon and all other exons for the same gene.

Details

Plot relative log2-fold-changes by exon for the specified gene. The relative logFC is the difference between the exon's logFC and the overall logFC for the gene, as computed by diffSplice. The significantly spliced individual exons are highlighted as red dots. The size of the red dots are weighted by its significance.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth and Yifang Hu

See Also

diffSplice, topSplice

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples

# See diffSplice

[Package limma version 3.40.6 Index]