plotWithHighlights {limma}	R Documentation

Scatterplot With Highlighting of Special Points

Description

Creates scatterplot, with optional size and color coding for points of special interest. This is the engine for plotMD and plotMA.

Usage


plotWithHighlights(x, y, status = NULL, values = NULL,
                   hl.pch = 16, hl.col = NULL, hl.cex = 1, legend = "topright",
                   bg.pch = 16, bg.col = "black", bg.cex = 0.3,
                   pch = NULL, col = NULL, cex = NULL, ...)

Arguments

x
numeric vector.

y
numeric vector of same length as x.

status
character vector giving the control status of each point, of same length as x and y. If NULL, then all points are plotted in the background color, symbol and size.

values
character vector giving values of status to be highlighted on the plot. Defaults to unique values of status in decreasing order of frequency, with the most frequent value set as the background value. Ignored if there is no status vector.

hl.pch
vector of plotting characters for highlighted points, either of unit length or of same length as values. Ignored is there is no status vector.

hl.col
vector of colors for highlighted points, either of unit length or of same length as values. Defaults to 1+1:length(values). Ignored if there is no status vector.

hl.cex
numeric vector of plot symbol expansions for highlighted points, either of unit length or of same length as values. Ignored if there is no status vector.

legend
character string giving position to place legend. See legend for possible values. Can also be logical, with FALSE meaning no legend. Ignored if there is no status vector.

bg.pch
plotting character for background (non-highlighted) points.

bg.col
color for background (non-highlighted) points.

bg.cex
plot symbol expansion for background (non-highlighted) points.

pch
synonym for hl.pch allowed for backward compatibility.

col
synonym for hl.col allowed for backward compatibility.

cex
synonym for hl.cex allowed for backward compatibility.

...
other arguments are passed to plot.

Details

This function produces a scatterplot in which the highlighted points are, by default, larger and colored compared to background points.

The status vector establishes the status of each point and values indicates which values of status should be highlighted. If values=NULL, then the most common value of status is assumed to correspond to background points and all other values are highlighted.

The arguments hl.pch, hl.col and hl.cex give graphics settings for highlighted points. By default, highlighted points are larger than background points and a different color is used for each distinct highlighted value.

The arguments bg.pch, bg.col and bg.cex give the graphics settings for non-highlighted (background) points. The same settings are used for all background points.

The arguments values, pch, col and cex can be included as attributes to status instead of being passed as arguments to plotWithHighlights. This is for compatibility with controlStatus.

See points for possible values for the graphics parameters.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

References

Ritchie, ME, Phipson, B, Wu, D, Hu, Y, Law, CW, Shi, W, and Smyth, GK (2015). limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic Acids Research 43, e47. http://nar.oxfordjournals.org/content/43/7/e47

See Also

plotMD, plotMA, mdplot

An overview of diagnostic plots available in LIMMA is given in 09.Diagnostics.

Examples


x <- runif(1000, min=4, max=16)
status <- rep(c(0,-1,1), c(950,40,10))
y <- status + rnorm(1000, sd=0.2)
plotWithHighlights(x, y, status=status)

[Package limma version 3.40.6 Index]