printorder {limma}	R Documentation

Identify Order in which Spots were Printed

Description

Identify order in which spots were printed and the 384-well plate from which they were printed.

Usage


printorder(layout, ndups=1, spacing="columns", npins, start="topleft")

Arguments

layout
list with the components ngrid.r, ngrid.c, nspot.r and nspot.c, or an RGList or MAList object from which the printer layout may be extracted.

ndups
number of duplicate spots, i.e., number of times print-head dips into each well

spacing
character string indicating layout of duplicate spots. Choices are "columns", "rows" or "topbottom".

npins
actual number of pins or tips on the print-head

start
character string giving position of the spot printed first in each grid. Choices are "topleft" or "topright" and partial matches are accepted.

Details

In most cases the printer-head contains the layout$ngrid.r times layout$ngrid.c pins or tips and the array is printed using layout$nspot.r times layout$npot.c dips of the head. The plate holding the DNA to be printed is assumed to have 384 wells in 16 rows and 24 columns.

ndups indicates the number of spots printed from each well. The replicate spots from multiple dips into the same wells are assumed to be side-by-side by columns (spacing="columns"), by rows (spacing="rows") or in the top and bottom halves of the array (spacing="topbottom").

In some cases a smaller number of physical pins is used and the total number of grids is built up by effectively printing two or more sub-arrays on the same slide. In this case the number of grids should be a multiple of the number of pins.

Printing is assumed to proceed by rows within in each grid starting either from the top-left or the top-right.

Value

List with components

printorder
numeric vector giving printorder of each spot, i.e., which dip of the print-head was used to print it

plate
numeric vector giving plate number from which each spot was printed

plate.r
numeric vector giving plate-row number of the well from which each spot was printed

plate.c
numeric vector giving plate-column number of the well from which each spot was printed

plateposition
character vector summarizing plate number and plate position of the well from which each spot was printed with letters for plate rows and number for columns. For example 02B13 is second row, 13th column, of the second plate.

Author(s)

Gordon Smyth

See Also

normalizeForPrintorder.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


printorder(list(ngrid.r=2,ngrid.c=2,nspot.r=12,nspot.c=8))

[Package limma version 3.40.6 Index]