qqt {limma}	R Documentation

Student's t or Fisher's F Quantile-Quantile Plot

Description

Plots the quantiles of a data sample against the theoretical quantiles of a Student's t distribution.

Usage

qqt(y, df = Inf, ylim = range(y), main = "Student's t Q-Q Plot", 
    xlab = "Theoretical Quantiles", ylab = "Sample Quantiles", plot.it = TRUE, ...) 
qqf(y, df1, df2, ylim=range(y), main= "F Distribution Q-Q Plot",
    xlab = "Theoretical Quantiles", ylab = "Sample Quantiles", plot.it = TRUE, ...)

Arguments

y
a numeric vector or array containing the data sample

df
degrees of freedom for the t-distribution. The default df=Inf represents the normal distribution.

df1
numerator degrees of freedom for the F-distribution.

df2
denominator degrees of freedom for the F-distribution.

ylim
plotting range for y

main
main title for the plot

xlab
x-axis title for the plot

ylab
y-axis title for the plot

plot.it
whether or not to produce a plot

...
other arguments to be passed to plot

Details

This function is analogous to qqnorm for normal probability plots. In fact qqt(y,df=Inf) is identical to qqnorm(y) in all respects except the default title on the plot.

Value

A list is invisibly returned containing the values plotted in the QQ-plot:

x
theoretical quantiles of the t-distribution or F-distribution

y
the data sample, same as input y

Author(s)

Gordon Smyth

See Also

qqnorm

Examples


#  See also the lmFit examples

y <- rt(50,df=4)
qqt(y,df=4)
abline(0,1)

[Package limma version 3.40.6 Index]