read.ilmn.targets {limma}	R Documentation

Read Illumina Data from a Target Dataframe

Description

Read Illumina data from a target dataframe

Usage


read.ilmn.targets(targets, ...)

Arguments

targets
data frame including names of profile files.

...
any other parameters are passed on to read.ilmn.

Details

targets is often created by calling the function readTargets. Rows in targets are arrays and columns contain related array or RNA sample information.

At least one of the two columns called files and/or ctrlfiles should be present in targets, which includes names of summary probe profile files and names of summary control probe profile files respectively. This function calls read.ilmn to read in the data.

Value

An EListRaw-class object. See return value of the function read.ilmn for details.

Author(s)

Wei Shi

See Also

read.ilmn

[Package limma version 3.40.6 Index]