read.matrix {limma}	R Documentation

Read Matrix with Headers from File

Description

Read a numeric matrix from a file assuming column headings on the first line. Not normally used directly by users.

This function is deprecated in limma 2.9.7 (1 January 2007).

Usage


read.matrix(file,nrows=0,skip=0,...)

Arguments

file 	 character string giving the file name

nrows 	 maximum number of rows of data to read, if greater than zero

skip 	 number of lines of the data file to skip before reading data

... 	 any other arguments to be passed to scan

Details

This function is similar to but faster than read.table(file,header=TRUE) when all the columns are numeric.

Value

A numeric matrix with column names.

Author(s)

Gordon Smyth

See Also

read.table, scan.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

[Package limma version 2.14.7 Index]