read.series {limma}	R Documentation

Read series of image files

Description

Read in a series of array image analysis output files as data frames.

This function is deprecated in limma 2.9.7 (1 January 2007).

Usage


read.series(slides, path=NULL, suffix="spot", ...)

Arguments

slides 	 character vector giving the names of files to be read in. Any suffix such as ".spot" or ".gpr" which is common to all the files can be omitted.

path 	 character string giving the directory containing the files. Can be omitted if the files are in the current working directory.

suffix 	 character string giving a suffix such as "spot" or "gpr" to be added to each file. If NULL then no suffix is added.

... 	 any other arguments to be passed to read.table

Details

This function performs a series of calls to read.table. The image analysis output files are assumed to have been edited to remove all pre-heading material. The files are assumed to contain only column names and data. In most cases only can use read.maimages instead.

The data.frames produced by this command will typically be processed further using one of the functions rg.spot, rg.genepix or rg.quantarray.

Value

No value is returned. However a series of data.frames are created on the current environment with names of the form filename.suffix. The files names are given by the elements of slides and the suffix is given by suffix.

Author(s)

Gordon Smyth

See Also

read.table.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

[Package limma version 2.14.7 Index]