readImaGeneHeader {limma}	R Documentation

Read ImaGene Header Information

Description

Read the header information from an ImaGene image analysis output file. This function is used internally by read.maimages and is not usually called directly by users.

Usage


readImaGeneHeader(file)

Arguments

file
character string giving file name or path

Details

The raw data files exported by the image analysis software ImaGene include a number of header lines which contain information about the printing and scanning processes. This function extracts that information and locates the line where the intensity data begins.

Value

A list containing information read from the header of the ImaGene file. Each Begin-End environment found in the file header will become a recursive list in the output object, with components corresponding to fields in the file. See the ImaGene documentation for further information. The output object will also contain a component NHeaderRecordsgiving the number of lines in the file before the intensity data begins.

Author(s)

Gordon Smyth

References

http://www.biodiscovery.com/software/imagene

See Also

read.imagene

An overview of LIMMA functions to read data is given in 03.ReadingData.

Examples

## Not run: 
h <- readImaGeneHeader("myImaGenefile.txt")
names(h)
h$NHeaderRecords
h[["Field Dimensions"]]

## End(Not run)

[Package limma version 3.40.6 Index]