readTargets {limma}	R Documentation

Read Targets File

Description

Read targets file for a microarray experiment into a dataframe.

Usage


readTargets(file="Targets.txt", path=NULL, sep="\t", row.names=NULL, quote="\"",...)

Arguments

file
character string giving the name of the targets file.

path
character string giving the directory containing the file. Can be omitted if the file is in the current working irectory.

sep
field separator character

row.names
character string giving the name of a column from which to obtain row names

quote
the set of quoting characters

...
other arguments are passed to read.table

Details

The targets file is a text file containing information about the RNA samples used as targets in the microarray experiment. Rows correspond to arrays and columns to covariates associated with the targets. For a two-color experiment, the targets file will normally include columns labelled Cy3 and Cy5 or similar specifying which RNA samples are hybridized to each channel of each array. Other columns may contain any other covariate information associated with the arrays or targets used in the experiment.

If row.names is non-null and there is a column by that name with unique values, then those values will be used as row names for the dataframe. If row.names is null, then the column Label will be used if such exists or, failing that, the column FileName.

See the Limma User's Guide for examples of this function.

Value

A dataframe. Character columns are not converted into factors.

Author(s)

Gordon Smyth

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

[Package limma version 3.40.6 Index]