readGAL {limma}	R Documentation

Read a GAL file

Description

Read a GenePix Array List (GAL) file into a dataframe.

Usage


readGAL(galfile=NULL,path=NULL,header=TRUE,sep="\t",quote="\"",skip=NULL,as.is=TRUE,...)

Arguments

galfile
character string giving the name of the GAL file. If NULL then a file with extension .gal is found in the directory specified by path.

path
character string giving the directory containing the files. If NULL then assumed to be the current working directory.

header
logical variable, if TRUE then the first line after skip is assumed to contain column headings. If FALSE then a value should specified for skip.

sep
the field separator character

quote
the set of quoting characters

skip
number of lines of the GAL file to skip before reading data. If NULL then this number is determined by searching the file for column headings.

as.is
logical variable, if TRUE then read in character columns as vectors rather than factors.

...
any other arguments are passed to read.table

Details

A GAL file is a list of genes IDs and associated information produced by an Axon microarray scanner. Apart from header information, the file must contain data columns labeled Block, Column, Row and ID. A Name column is usually included as well. Other columns are optional. See the Axon URL below for a detaile description of the GAL file format.

This function reads in the data columns with a minimum of user information. In most cases the function can be used without specifying any of the arguments.

Value

A data frame with columns

Block
numeric vector containing the print tip indices

Column
numeric vector containing the spot columns

Row
numeric vector containing the spot rows

ID
character vector, for factor if as.is=FALSE, containing gene library identifiers

Name
character vector, for factor if as.is=FALSE, containing gene names

The data frame will be sorted so that Column is the fastest moving index, then Row, then Block.

Author(s)

Gordon Smyth

References

http://www.cryer.co.uk/file-types/a/atf/genepix_file_formats.htm

See Also

read.Galfile in the marray package.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


# readGAL()
# will read in the first GAL file (with suffix ".gal")
# found in the current working directory

[Package limma version 3.40.6 Index]