residuals.MArrayLM {limma}	R Documentation

Extract Residuals from MArrayLM Fit

Description

This method extracts the residuals from all the probewise linear model fits and returns them in a matrix.

Usage


## S3 method for class 'MArrayLM'
residuals(object, y, ...)

Arguments

object
a fitted model object inheriting from class MarrayLM.

y
a data object containing the response data used to compute the fit. This can be of any class for which as.matrix is defined, including MAList, ExpressionSet, marrayNorm etc.

...
other arguments are not used

Value

Numeric matrix of residuals.

See Also

residuals.

[Package limma version 3.40.6 Index]