rg.genepix {limma}	R Documentation

Extract RGList from data.frames Containing GenePix Data

Description

Extracts an RGList from GenePix image analysis output when the data has already been read from files into data.frames objects.

This function is deprecated in limma 2.9.7 (1 January 2007).

Usage


rg.genepix(slides,names.slides=names(slides),suffix="gpr")

Arguments

slides 	 character vector giving the names of the data frames containing the Spot output.

names.slides 	 character vector giving column names to be associated with each slide.

suffix 	 the dataframe names are assumed to have this suffix added to names in slides.

Details

This function extracts the foreground and background intensities from a series of data frames and assembles them in the components of one list.

Value

A list containing the components

R 	 A matrix containing the red channel foreground intensities for each spot for each array.

Rb 	 A matrix containing the red channel background intensities for each spot for each array.

G 	 A matrix containing the green channel foreground intensities for each spot for each array.

Gb 	 A matrix containing the green channel background intensities for each spot for each array.

Author(s)

Gordon Smyth

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

[Package limma version 2.14.7 Index]