rg.series.spot {limma}	R Documentation

Read RGList from SPOT Image Analysis Output Files

Description

Extracts an RGList from a series of Spot image analysis files.

This is a depreciated function. Use read.maimages instead.

Usage


rg.series.spot(slides,path=NULL,names.slides=names(slides),suffix="spot",wt.fun=NULL,verbose=TRUE,...)

Arguments

slides 	 character vector giving the names of the files containing the Spot output

path 	 character string giving the directory containing the files. Can be omitted if the files are in the current working directory.

names.slides 	 names to be associated with each slide as column name

suffix 	 the file names are assumed to have this suffix added to names in slides

wt.fun 	 function to calculate quality weights

verbose 	 TRUE to report each time a file in read in

... 	 any other arguments to be passed to scan

Details

This function extracts the foreground and background intensities from a series of files, produced by the image analysis program SPOT, and assembles them in the components of one list.

Spot quality weights may also be extracted using an optional weight function.

Value

A list containing the components

R 	 matrix containing the red channel foreground intensities for each spot for each array.

Rb 	 matrix containing the red channel background intensities for each spot for each array.

G 	 matrix containing the green channel foreground intensities for each spot for each array.

Gb 	 matrix containing the green channel background intensities for each spot for each array.

weights 	 spot quality weights, if wt.fun is not NULL

Author(s)

Gordon Smyth

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

[Package limma version 2.14.7 Index]