subsetting {limma}	R Documentation

Subset RGList, MAList, EListRaw, EList or MArrayLM Objects

Description

Return an RGList, MAList, EListRaw, EList or MArrayLM object with only selected rows and columns of the original object.

Usage


## S3 method for class 'RGList'
object[i, j]
subsetListOfArrays(object, i, j, IJ, IX, I, JX)

Arguments

object
object of class RGList, MAList, EListRaw, EList or MArrayLM.

i,j
elements to extract. i subsets the probes or spots while j subsets the arrays.

IJ
character vector giving names of components that should be subsetted by i and j.

IX
character vector giving names of 2-dimensional components that should be subsetted by i only.

I
character vector giving names of vector components that should be subsetted by i.

JX
character vector giving names of 2-dimensional components whose row dimension corresponds to j.

Details

i,j may take any values acceptable for the matrix components of object. Either or both can be missing. See the Extract help entry for more details on subsetting matrices.

object[] will return the whole object unchanged. A single index object[i] will be taken to subset rows, so object[i] and object[i,] are equivalent.

subsetListOfArrays is used internally as a utility function by the subsetting operations. It is not intended to be called directly by users. Values must be supplied for all arguments other than i and j.

Value

An object the same as object but containing data from the specified subset of rows and columns only.

Author(s)

Gordon Smyth

See Also

Extract in the base package.

02.Classes for a summary of the different data classes.

Examples


M <- A <- matrix(11:14,4,2)
rownames(M) <- rownames(A) <- c("a","b","c","d")
colnames(M) <- colnames(A) <- c("A","B")
MA <- new("MAList",list(M=M,A=A))
MA[1:2,]
MA[c("a","b"),]
MA[1:2,2]
MA[,2]

[Package limma version 3.40.6 Index]