topSplice {limma}	R Documentation

Top table of differentially spliced genes or exons

Description

Top table ranking the most differentially spliced genes or exons.

Usage


topSplice(fit, coef = ncol(fit), test = "simes", number = 10, FDR=1, sort.by = "p")

Arguments

fit
MArrayLM fit object produced by diffSplice.

coef
the coefficient (column) of fit for which differentially splicing is assessed.

test
character string specifying which statistical test to apply. Possible values are "simes", "F" or "t". "F" gives F-tests for each gene. "t" gives t-tests for each exon. "simes" gives genewise p-values derived from the t-tests after Simes adjustment for each gene.

number
integer, maximum number of rows to output.

FDR
numeric, only show exons or genes with false discovery rate less than this cutoff.

sort.by
character string specifying which column to sort results by. Possible values for "p", "logFC", "NExons" or "none". "logFC" is only available if test="t" and "NExons" is only available if test="simes" or test="F".

Details

Ranks genes or exons by evidence for differential splicing. The F-statistic tests for any differences in exon usage between experimental conditions. The exon-level t-statistics test for differences between each exon and all other exons for the same gene.

The Simes processes the exon-level p-values to give an overall call of differential splicing for each gene. It returns the minimum Simes-adjusted p-values for each gene.

The F-tests are likely to be powerful for genes in which several exons are differentially splices. The Simes p-values is likely to be more powerful when only a minority of the exons for a gene are differentially spliced. The exon-level t-tests are not recommended for formal error rate control.

Value

A data.frame with any annotation columns found in fit plus the following columns

logFC
log2-fold change of exon vs other exons for the same gene (if level="exon")

t
moderated t-statistic (if level="exon")

F
moderated F-statistic (if level="gene")

P.Value
p-value

FDR
false discovery rate

Author(s)

Gordon Smyth

See Also

diffSplice, plotSplice

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples

# See diffSplice

[Package limma version 3.40.6 Index]