trigammaInverse {limma}	R Documentation

Inverse Trigamma Function

Description

The inverse of the trigamma function.

Usage


trigammaInverse(x)

Arguments

x
numeric vector or array

Details

The function uses Newton's method with a clever starting value to ensure monotonic convergence.

Value

Numeric vector or array y satisfying trigamma(y)==x.

Note

This function does not accept a data.frame as argument although the base package function trigamma does.

Author(s)

Gordon Smyth

See Also

This function is the inverse of trigamma in the base package.

This function is called by fitFDist.

Examples


y <- trigammaInverse(5)
trigamma(y)

[Package limma version 3.40.6 Index]