trimWhiteSpace {limma}	R Documentation

Trim Leading and Trailing White Space

Description

Trims leading and trailing white space from character strings.

Usage

trimWhiteSpace(x)

Arguments

x
character vector

Value

A character vector of the same length as x in which leading and trailing white space has been stripped off each value.

Author(s)

Tim Beissbarth and Gordon Smyth

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples

x <- c("a "," b ")
trimWhiteSpace(x)

[Package limma version 3.66.0 Index]