weighted.median {limma}	R Documentation

Weighted Median

Description

Compute a weighted median of a numeric vector.

Usage


weighted.median(x, w, na.rm = FALSE)

Arguments

x
a numeric vector containing the values whose mean is to be computed.

w
a vector of weights the same length as x giving the weights to use for each element of x.

na.rm
a logical value indicating whether NA values in x should be stripped before the computation proceeds.

Details

If w is missing then all elements of x are given the same weight.

Missing values in w are not handled.

The weighted median is the median of the discrete distribution with values given by x and probabilities given by w/sum(w).

Value

numeric value giving the weighted median

See Also

median, weighted.mean

Examples


## GPA from Siegel 1994
wt <- c(5,  5,  4,  1)/15
x <- c(3.7,3.3,3.5,2.8)
xm <- weighted.median(x,wt)

[Package limma version 3.40.6 Index]