wsva {limma}	R Documentation

Weighted Surrogate Variable Analysis

Description

Calculate surrogate variables from the singular vectors of the linear model residual space.

Usage


wsva(y, design, n.sv = 1L, weight.by.sd = FALSE, plot = FALSE, ...)

Arguments

y
numeric matrix giving log-expression or log-ratio values for a series of microarrays, or any object that can coerced to a matrix including ExpressionSet, MAList, EList or PLMSet objects. Rows correspond to genes and columns to samples.

design
design matrix

n.sv
number of surrogate variables required.

weight.by.sd
logical, should the surrogate variables be especially tuned to the more variable genes?

plot
logical. If TRUE, plots the proportion of variance explained by each surrogate variable.

...
other arguments can be included that would be suitable for lmFit.

Details

The function constructs surrogate variables that explain a high proportion of the residual variability for many of the genes. The surrogate variables can be included in the design matrix to remove unwanted variation. The surrogate variables are constructed from the singular vectors of a representation of the linear model residual space.

If weight.by.sd=FALSE, then the method is a simplification of the approach by Leek and Storey (2007).

Value

Numeric matrix with ncol(y) rows and n.sv columns containing the surrogate variables.

Author(s)

Gordon Smyth and Yifang Hu

References

Leek, JT, Storey, JD (2007). Capturing heterogeneity in gene expression studies by surrogate variable analysis. PLoS Genetics 3, 1724-1735.

[Package limma version 3.40.6 Index]