04.Background {limma}	R Documentation

Background Correction

Description

This page deals with background correction methods for two-color microarray data.

Usually one doesn't need to explicitly ask for background correction of the intensities because this is done by default by normalizeWithinArrays, which subtracts the background from the foreground intensities before applying the normalization method. This default background correction method can be over-ridden by using backgroundCorrect which offers a number of alternative background correct methods to simple subtraction. The function backgroundCorrect is used to correct the RGList before applying normalizeWithinArrays.

The movingmin method of backgroundCorrect uses utility functions ma3x3.matrix and ma3x3.spottedarray.

The normexp method of backgroundCorrect uses ulility functions normexp.fit, normexp.signal, normexp.m2loglik, normexp.grad, dnormexp and dnormexp.saddle.

kooperberg is a Bayesian background correction tool designed specifically for GenePix data. kooperberg is not currently used as the default method for GenePix data because it is computationally intensive. It requires several additional columns from the GenePix data files which can be read in using read.maimages and specifying the other.columns argument.

Author(s)

Gordon Smyth

[Package limma version 2.14.7 Index]