05.Normalization {limma}	R Documentation

Normalization of Microarray Data

Description

This page gives an overview of the LIMMA functions available to normalize data from spotted two-colour microarrays. Smyth and Speed (2003) give an overview of the normalization techniques implemented in the functions.

Usually data from spotted microarrays will be normalized using normalizeWithinArrays. A minority of data will also be normalized using normalizeBetweenArrays if diagnostic plots suggest a difference in scale between the arrays.

In rare circumstances, data might be normalized using normalizeForPrintorder before using normalizeWithinArrays.

All the normalization routines take account of spot quality weights which might be set in the data objects. The weights can be temporarily modified using modifyWeights to, for example, remove ratio control spots from the normalization process.

If one is planning analysis of single-channel information from the microarrays rather than analysis of differential expression based on log-ratios, then the data should be normalized using a single channel-normalization technique. Single channel normalization uses further options of the normalizeBetweenArrays function. For more details see the LIMMA User's Guide which includes a section on single-channel normalization.

normalizeWithinArrays uses utility functions MA.RG, loessFit and normalizeRobustSpline. normalizeBetweenArrays uses utility functions normalizeMedianAbsValues and normalizeQuantiles, none of which need to be called directly by users.

Author(s)

Gordon Smyth

References

Smyth, G. K., and Speed, T. P. (2003). Normalization of cDNA microarray data. In: METHODS: Selecting Candidate Genes from DNA Array Screens: Application to Neuroscience, D. Carter (ed.). Methods Volume 31, Issue 4, December 2003, pages 265-273. http://www.statsci.org/smyth/pubs/normalize.pdf

[Package limma version 2.14.7 Index]