09.Diagnostics {limma}	R Documentation

Diagnostics and Quality Assessment

Description

This page gives an overview of the LIMMA functions available for microarray quality assessment and diagnostic plots.

This package provides an anova method which is designed for assessing the quality of an array series or of a normalization method. It is not designed to assess differential expression of individual genes. anova uses utility functions bwss and bwss.matrix.

The function arrayWeights estimates the empirical reliability of each array following a linear model fit.

Diagnostic plots can be produced by

imageplot

Produces a spatial picture of any spot-specific measure from an array image. If the log-ratios are plotted, then this produces an in-silico representation of the well known false-color TIFF image of an array. imageplot3by2 will write imageplots to files, six plots to a page.

plotFB

Plots foreground versus background log-intensies for an array.

plotMA

MA-plots. One of the most useful plots of a two-color array. plotMA3by2 will write MA-plots to files, six plots to a page. mdplot can also be useful for comparing two one-channel microarrays.

plotPrintTipLoess

Produces a grid of MA-plots, one for each print-tip group on an array, together with the corresponding lowess curve. Intended to help visualize print-tip loess normalization.

plotPrintorder

For an array, produces a scatter plot of log-ratios or log-intensities by print order.

plotDensities

Individual channel densities for one or more arrays. An essential plot to accompany between array normalization, especially quantile normalization.

plotPrintTipLoess uses utility functions gridr and gridc. plotDensities uses utility function RG.MA.

Author(s)

Gordon Smyth

[Package limma version 2.14.7 Index]