arrayWeightsQuick {limma}	R Documentation

Array Quality Weights

Description

Estimates relative quality weights for each array in a multi-array experiment with replication.

Usage


arrayWeightsQuick(y, fit)

Arguments

y 	 the data object used to estimate fit. Can be of any class which can be coerced to matrix, including matrix, MAList, marrayNorm or ExpressionSet.

fit 	 MArrayLM fitted model object

Details

Estimates the relative reliability of each array by measuring how well the expression values for that array follow the linear model.

This is a quick and dirty version of arrayWeights.

Value

Numeric vector of weights of length ncol(fit).

Author(s)

Gordon Smyth

References

Ritchie, M. E., Diyagama, D., Neilson, van Laar, R., J., Dobrovic, A., Holloway, A., and Smyth, G. K. (2006). Empirical array quality weights for microarray data. BMC Bioinformatics. (Accepted 11 April 2006)

See Also

See arrayWeights. An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


## Not run: 
fit <- lmFit(y, design)
arrayWeightsQuick(y, fit)
## End(Not run)

[Package limma version 2.14.7 Index]