as.data.frame {limma}	R Documentation

Turn a Microarray Linear Model Object into a Dataframe

Description

Turn a MArrayLM object into a data.frame.

Usage

## S3 method for class 'MArrayLM'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

Arguments

x
an object of class MArrayLM

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, setting row names and converting column names (to syntactic names) is optional.

...
additional arguments to be passed to or from methods.

Details

This method combines all the components of x which have a row for each probe on the array into a data.frame.

Value

A data.frame.

Author(s)

Gordon Smyth

See Also

as.data.frame in the base package.

02.Classes gives an overview of data classes used in LIMMA. 06.LinearModels gives an overview of linear model functions in LIMMA.

[Package limma version 3.66.0 Index]