as.MAList {limma}	R Documentation

Convert marrayNorm Object to an MAList Object

Description

Convert marrayNorm Object to an MAList Object

Usage

as.MAList(object)

Arguments

object
an marrayNorm object

Details

The marrayNorm class is defined in the marray package. This function converts a normalized two color microarray data object created by the marray package into the corresponding limma data object.

Note that such conversion is not necessary to access the limma linear modelling functions, because lmFit will operate on a marrayNorm data object directly.

Value

Object of class MAList

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of all the classes defined by this package.

The marrayNorm class is defined in the marray package.

[Package limma version 3.66.0 Index]