as.matrix {limma}	R Documentation

Turn a Microarray Data Object into a Matrix

Description

Turn a microarray data object into a numeric matrix by extracting the expression values.

Usage


## S3 method for class 'MAList':
as.matrix(x,...)

Arguments

x 	 an object of class RGList, MAList, MArrayLM, marrayNorm, PLMset, ExpressionSet, LumiBatch or vsn.

... 	 additional arguments, not used for these methods.

Details

These methods extract the matrix of log-ratios, for MAList or marrayNorm objects, or the matrix of expression values for other expression objects such as ExressionSet. For MArrayLM objects, the matrix of fitted coefficients is extracted.

These methods involve loss of information, so the original data object is not recoverable.

Value

A numeric matrix.

Author(s)

Gordon Smyth

See Also

as.matrix in the base package or exprs in the Biobase package.

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 2.14.7 Index]