avedups {limma}	R Documentation

Average Over Duplicate Spots

Description

Condense a microarray data object so that values for within-array replicate spots are replaced with their average.

Usage

## Default S3 method:
avedups(x, ndups=2, spacing=1, weights=NULL)
## S3 method for class 'MAList'
avedups(x, ndups=x$printer$ndups, spacing=x$printer$spacing, weights=x$weights)
## S3 method for class 'EList'
avedups(x, ndups=x$printer$ndups, spacing=x$printer$spacing, weights=x$weights)

Arguments

x
a matrix-like object, usually a matrix, MAList or EList object.

ndups
number of within-array replicates for each probe.

spacing
number of spots to step from a probe to its duplicate.

weights
numeric matrix of spot weights.

Details

A new data object is computed in which each probe is represented by the (weighted) average of its duplicate spots. For an MAList object, the components M and A are both averaged in this way. For an EList object, the component E is averaged in this way.

If x is of mode "character", then the duplicate values are assumed to be equal and the first is taken as the average.

Value

A data object of the same class as x with 1/ndups as many rows.

Author(s)

Gordon Smyth

See Also

avereps.

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 3.66.0 Index]