avereps {limma}	R Documentation

Average Over Irregular Replicate Spots

Description

Condense a microarray data object so that values for within-array replicate spots are replaced with their average.

Usage


## Default S3 method:
avereps(x, ID=rownames(x))
## S3 method for class 'MAList':
avereps(x, ID=NULL)

Arguments

x 	 a matrix-like object, usually a matrix or an MAList object.

ID 	 probe identifier.

Details

A new data object is computed in which each probe is represented by the average of its replicate spots. For an MAList object, the components M and A are both averaged in this way, as weights and any matrices found in object$other.

For an MAList object, ID defaults to MA$genes$ID is that exists, otherwise to rownames(MA$M).

If x is of mode "character", then the replicate values are assumed to be equal and the first is taken as the average.

Value

A data object of the same class as x with a row for each unique value of ID.

Author(s)

Gordon Smyth

See Also

dimnames in the base package.

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 2.14.7 Index]