cbind {limma}	R Documentation

Combine RGList or MAList Objects

Description

Combine a series of RGList objects or combine a series of MAList objects.

Usage


## S3 method for class 'RGList':
cbind(..., deparse.level=1)
## S3 method for class 'RGList':
rbind(..., deparse.level=1)

Arguments

... 	 RGList objects or MAList objects

deparse.level 	 not currently used, see cbind in the base package

Details

cbind combines data objects assuming the same gene lists but different arrays. rbind combines data objects assuming equivalent arrays, i.e., the same RNA targets, but different genes.

For cbind, the matrices of expression data from the individual objects are cbinded. The data.frames of target information, if they exist, are rbinded. The combined data object will preserve any additional components or attributes found in the first object to be combined. For rbind, the matrices of expression data are rbinded while the target information, in any, is unchanged.

Value

An RGList or MAList object holding data from all the arrays and all genes from the individual objects.

Author(s)

Gordon Smyth

See Also

cbind in the base package.

03.ReadingData gives an overview of data input and manipulation functions in LIMMA.

Examples


M <- A <- matrix(11:14,4,2)
rownames(M) <- rownames(A) <- c("a","b","c","d")
colnames(M) <- colnames(A) <- c("A1","A2")
MA1 <- new("MAList",list(M=M,A=A))

M <- A <- matrix(21:24,4,2)
rownames(M) <- rownames(A) <- c("a","b","c","d")
colnames(M) <- colnames(A) <- c("B1","B2")
MA2 <- new("MAList",list(M=M,A=A))

cbind(MA1,MA2)

[Package limma version 2.14.7 Index]