changeLog {limma}	R Documentation

Change Log

Description

Show the most recent changes from a package change log or NEWS file.

Usage

changeLog(n = 30, package = "limma")

Arguments

n
integer, number of lines to write of change log.

package
character string giving name of package.

Details

The function will look for a file changelog.txt or ChangeLog in the top-level or doc directories of the installed package. Failing that, it will look for NEWS or NEWS.md in the top-level directory.

Note that changeLog does not write the content of NEWS.Rd, which is a structured file. Use news(package="limma") for that instead.

Value

No value is produced, but a number of lines of text are written to standard output.

Author(s)

Gordon Smyth

See Also

01.Introduction, news.

Examples

changeLog()
changeLog(package="statmod")

[Package limma version 3.50.3 Index]