cumOverlap {limma}	R Documentation

Cumulative Overlap Analysis of Ordered Lists

Description

Test whether the leading members of ordered lists significantly overlap.

Usage

cumOverlap(ol1, ol2)

Arguments

ol1
vector containing first ordered list.

ol2
vector containing second ordered list.

Details

The function compares the top n members of each list, for every possible n, and conducts an hypergeometric test for overlap. The function returns the value of n giving the smallest Bonferroni adjusted p-value.

This method was described in Chapter 4 of Wu (2011).

Value

List containing the following components:

n.min
integer, top table length leading to smallest adjusted p-value.

p.min
smallest adjusted p-value.

n.overlap
integer, number of overlapping IDs in first n.min.

id.overlap
vector giving the overlapping IDs in first n.min.

p.value
numeric, vector of p-values for each possible top table length.

adj.p.value
numeric, vector of Bonferroni adjusted p-values for each possible top table length.

Author(s)

Gordon Smyth and Di Wu

References

Wu, D (2011). Finding hidden relationships between gene expression profiles with application to breast cancer biology. PhD thesis, University of Melbourne. http://hdl.handle.net/11343/36278

Examples

ol1 <- letters[1:26]
ol2 <- letters[sample(26)]
coa <- cumOverlap(ol1, ol2)
coa$p.min

[Package limma version 3.34.9 Index]