diffSplice {limma}	R Documentation

Test for Differential Transcript Usage

Description

Given a linear model fit at the transcript (or exon) level, test for differences in transcript (or exon) usage within genes between experimental conditions. More generally, test for differential usage within genes of any set of splice-events or isoform-identifying features.

Usage

## S3 method for class 'MArrayLM'
diffSplice(fit, geneid, exonid = NULL,
     robust = FALSE, legacy = FALSE, verbose = TRUE, ...)

Arguments

fit
an MArrayLM fitted model object produced by voomLmFit, lmFit or contrasts.fit. Rows should correspond to transcripts, for a DTU analysis, or to exons and exon-exon junctions for a DEU analysis.

geneid
gene identifiers. Either a vector of length nrow(fit) or the name of the column of fit$genes containing the gene identifiers. Rows with the same ID are assumed to belong to the same gene.

exonid
exon identifiers. Either a vector of length nrow(fit) or the name of the column of fit$genes containing the exon identifiers.

robust
logical, should the estimation of the empirical Bayes prior parameters be robustified against outlier sample variances?

legacy
logical. If FALSE then the new empirical Bayes hyperparameter estimation (introduced in limma 3.61.8) will be used, if TRUE the earlier hyperparameter estimation will be used. The new method is particularly appropriate when the residual degrees of freedom are not all equal, which is likely to be the case for diffSplice.

verbose
logical, if TRUE some diagnostic information about the number of genes and exons is output.

...
other arguments are not currently used.

Details

This function tests for differential usage of the row-wise isoform features contained in fit for each gene and for each column of fit. The isoform features can be transcripts for a differential transcript usage (DTU) analysis, or can be a combination of exons and exon-exon junctions for a differential exon usage (DEU) analysis.

Testing for differential transcript usage is equivalent to testing whether the log-fold-changes in the fit differ between transcripts for the same gene. Two different tests are provided. The first is a F-test for differences between the log-fold-changes for each gene. This is equivalent to testing for interaction between the transcripts for that gene and the coefficient of the linear model. The other is a series of t-tests in which each transcript is compared to the weighted average of all other transcripts for the same gene. The transcript-level t-tests are converted into genewise tests by adjusting the p-values for the same gene by Simes method. Alternatively, the transcript-level t-tests are also converted into genewise tests by adjusting the smallest p-value for each gene by Bonferroni's method.

This function can be used on transcript level RNA-seq counts from Salmon or kallisto, after using the edgeR functions catchSalmon() or catchKallisto() and voomLmFit(), as described by Baldoni et al (2025). It can also be used on equivalence-class counts from Salmon or kallisto, after pre-processing by voomLmFit(), as described by Cmero et al (2019). It can also be used on exon-level read counts or on data from an exon microarray.

Value

An object of class MArrayLM containing both exon level and gene level tests. Results are sorted by geneid and by exonid within gene.

coefficients
numeric matrix of coefficients of same dimensions as fit. Each coefficient is the difference between the log-fold-change for that exon versus the average log-fold-change for all other exons for the same gene.

t
numeric matrix of moderated t-statistics, of same dimensions as fit.

p.value
numeric vector of p-values corresponding to the t-statistics

genes
data.frame of exon annotation

genecolname
character string giving the name of the column of genes containing gene IDs

gene.F
numeric matrix of moderated F-statistics, one row for each gene.

gene.F.p.value
numeric matrix of p-values corresponding to gene.F

gene.simes.p.value
numeric matrix of Simes adjusted p-values, one row for each gene.

gene.bonferroni.p.value
numeric matrix of Bonferroni adjusted p-values, one row for each gene.

gene.genes
data.frame of gene annotation.

Note

This function is not designed for situations with a very high level of multi-counting of RNA-seq reads that overlap two or more exons. In particular, it is not designed for use with "chopped" exons, where overlapping exons belonging to different transcripts of the same gene are chopped up into unique sub-exons, because artificial exons of this sort lead to high levels of multi-counting.

Author(s)

Gordon Smyth and Charity Law

References

Baldoni PL, Chen L, Li M, Chen Y, Smyth GK (2025). Dividing out quantification uncertainty enables assessment of differential transcript usage with limma and edgeR. bioRxiv doi:10.1101/2025.04.07.647659.

Cmero M, Davidson NM, Oshlack A (2019). Using equivalence class counts for fast and accurate testing of differential transcript usage. F1000Research 8, 265. doi:10.12688/f1000research.18276.2.

See Also

topSplice and plotSplice are downstream functions that operate on the output from diffSplice.

Also see diffSplice.DGEGLM in the edgeR package, which has comparable functionality but for edgeR fit objects.

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples

## Not run: 
fit <- voomLmFit(dge, design)
ex <- diffSplice(fit, geneid="GeneID")
topSplice(ex)
plotSplice(ex, xlab="Transcript")

## End(Not run)

[Package limma version 3.66.0 Index]