fitFDist {limma}	R Documentation

Moment Estimation of Scaled F-Distribution

Description

Moment estimation of the parameters of a scaled F-distribution given one of the degrees of freedom. This function is called internally by ebayes and is not usually called directly by a user.

Usage


fitFDist(x,df1)

Arguments

x 	 numeric vector or array of positive values representing a sample from an F-distribution.

df1 	 the first degrees of freedom of the F-distribution. May be an integer or a vector of the same length as x.

Details

The function estimates scale and df2 under the assumption that x is distributed as scale times an F-distributed random variable on df1 and df2 degrees of freedom.

Value

A list containing the components

scale 	 scale factor for F-distribution

df2 	 the second degrees of freedom of the F-distribution

Author(s)

Gordon Smyth

See Also

ebayes, trigammaInverse

[Package limma version 2.14.7 Index]