getEAWP {limma}	R Documentation

Extract Basic Data from Microarray Data Objects

Description

Given a microarray data object of any known class, get the expression values, weights, probe annotation and A-values, which are needed for linear modelling. This function is called by the linear modelling functions in LIMMA.

Usage


getEAWP(object)

Arguments

object 	 a microarray data object. An object of class list, MAList, marrayNorm, PLMset, vsn, or any class inheriting from ExpressionSet, or any object that can be coerced to a numeric matrix.

Details

In the case of two-color objects, the Amean is computed from the matrix of A-values. For single-channel objects, Amean is computed from the matrix of expression vales. PLMset, vsn and ExpressionSet are assumed to be single-channel for this purpose.

If object is a matrix, it is assumed to contain log-intensities if the values are all positive and log-ratios otherwise. Amean is computed in the former case but not the latter.

Value

A list with components

exprs 	 numeric matrix of log-ratios or log-intensities

weights 	 numeric matrix of weights

probes 	 data.frame of probe-annotation

Amean 	 numeric vector of average log-expression for each probe

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 2.14.7 Index]