getSpacing {limma}	R Documentation

Get Numerical Spacing

Description

Convert character to numerical spacing measure for within-array replicate spots.

Usage


getSpacing(spacing, layout)

Arguments

spacing 	 character string or integer. Acceptable character strings are columns, rows or topbottom. Integer values are simply passed through.

layout 	 list containing printer layout information

Value

Integer giving spacing between replicate spots in the gene list.

Author(s)

Gordon Smyth

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


getSpacing("columns",list(ngrid.r=2,ngrid.c=2,nspot.r=20,nspot.c=19))
getSpacing("rows",list(ngrid.r=2,ngrid.c=2,nspot.r=20,nspot.c=19))
getSpacing("topbottom",list(ngrid.r=2,ngrid.c=2,nspot.r=20,nspot.c=19))

[Package limma version 2.14.7 Index]