getLayout {limma}	R Documentation

Extract the Print Layout of an Array from the GAL File

Description

From the Block, Row and Column information in a genelist, determine the number of grid rows and columns on the array and the number of spot rows and columns within each grid.

Usage


getLayout(gal, guessdups=FALSE)
getLayout2(galfile)

Arguments

gal 	 data.frame containing the GAL, i.e., giving the position and gene identifier of each spot

galfile 	 name or path of GAL file

guessdups 	 logical, if TRUE then try to determine number and spacing of duplicate spots, i.e., within-array replicates

Details

A GenePix Array List (GAL) file is a list of genes and associated information produced by an Axon microarray scanner. The function getLayout determines the print layout from a data frame created from a GAL file or gene list. The data.frame must contain columns Block, Column and Row. (The number of tip columns is not assumed to be either one or four.) The function getLayout2 determines the print layout from the header information of an actual GAL file.

Value

A printlayout object, which is a list with the following components. The last two components are present only if guessdups=TRUE.

ngrid.r 	 integer, number of grid rows on the arrays

ngrid.c 	 integer, number of grid columns on the arrays

nspot.r 	 integer, number of rows of spots in each grid

nspot.c 	 integer, number of columns of spots in each grid

ndups 	 integer, number of times each probe is printed on the array

spacing 	 integer, spacing between multiple printings of each probe

Author(s)

Gordon Smyth and James Wettenhall

See Also

gpTools.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


# gal <- readGAL()
# layout <- getLayout(gal)

[Package limma version 2.14.7 Index]