heatdiagram {limma}	R Documentation

Stemmed Heat Diagram

Description

Creates a heat diagram showing the co-regulation of genes under one condition with a range of other conditions.

Usage


heatDiagram(results,coef,primary=1,names=NULL,treatments=colnames(coef),limit=NULL,orientation="landscape",low="green",high="red",cex=1,mar=NULL,ncolors=123,...)
heatdiagram(stat,coef,primary=1,names=NULL,treatments=colnames(stat),critical.primary=4,critical.other=3,limit=NULL,orientation="landscape",low="green",high="red",cex=1,mar=NULL,ncolors=123,...)

Arguments

results 	 TestResults matrix, containing elements -1, 0 or 1, from decideTests

stat 	 numeric matrix of test statistics. Rows correspond to genes and columns to treatments or contrasts between treatments.

coef 	 numeric matrix of the same size as stat. Holds the coefficients to be displayed in the plot.

primary 	 number or name of the column to be compared to the others. Genes are included in the diagram according to this column of stat and are sorted according to this column of coef. If primary is a name, then stat and coef must have the same column names.

names 	 optional character vector of gene names

treatments 	 optional character vector of treatment names

critical.primary 	 critical value above which the test statistics for the primary column are considered significant and included in the plot

critical.other 	 critical value above which the other test statistics are considered significant. Should usually be no larger than critical.primary although larger values are permitted.

limit 	 optional value for coef above which values will be plotted in extreme color. Defaults to max(abs(coef)).

orientation 	 "portrait" for upright plot or "landscape" for plot orientated to be wider than high. "portrait" is likely to be appropriate for inclusion in printed document while "landscape" may be appropriate for a presentation on a computer screen.

low 	 color associated with repressed gene regulation

high 	 color associated with induced gene regulation

ncolors 	 number of distinct colors used for each of up and down regulation

cex 	 factor to increase or decrease size of column and row text

mar 	 numeric vector of length four giving the size of the margin widths. Default is cex*c(5,6,1,1) for landscape and cex*c(1,1,4,3) for portrait.

... 	 any other arguments will be passed to the image function

Details

Users are encouraged to use heatDiagram rather than heatdiagram as the later function may be removed in future versions of limma.

This function plots an image of gene expression profiles in which rows (or columns for portrait orientation) correspond to treatment conditions and columns (or rows) correspond to genes. Only genes which are significantly differentially expressed in the primary condition are included. Genes are sorted by differential expression under the primary condition.

Note: the plot produced by this function is unique to the limma package. It should not be confused with "heatmaps" often used to display results from cluster analyses.

Value

An image is created on the current graphics device. A matrix with named rows containing the coefficients used in the plot is also invisibly returned.

Author(s)

Gordon Smyth

See Also

image.

Examples


library(sma)
data(MouseArray)
MA <- normalizeWithinArrays(mouse.data,layout=mouse.setup)
design <- cbind(c(1,1,1,0,0,0),c(0,0,0,1,1,1))
fit <- lmFit(MA,design=design)
contrasts.mouse <- cbind(Control=c(1,0),Mutant=c(0,1),Difference=c(-1,1))
fit <- eBayes(contrasts.fit(fit,contrasts=contrasts.mouse))
results <- decideTests(fit,method="global",p=0.1)
heatDiagram(results,fit$coef,primary="Difference")

[Package limma version 2.14.7 Index]