helpMethods {limma}	R Documentation

Prompt for Method Help Topics

Description

For any S4 generic function, find all methods defined in currently loaded packages. Prompt the user to choose one of these to display the help document.

Usage


helpMethods(genericFunction)

Arguments

genericFunction 	 a generic function or a character string giving the name of a generic function

Author(s)

Gordon Smyth

See Also

showMethods

Examples


## Not run: helpMethods(show)

[Package limma version 2.14.7 Index]