is.fullrank {limma}	R Documentation

Check for Full Column Rank

Description

Test whether a numeric matrix has full column rank.

Usage


is.fullrank(x)
nonEstimable(x)

Arguments

x 	 a numeric matrix or vector

Details

is.fullrank is used to check the integrity of design matrices in limma, for example after subsetting operations.

nonEstimable is used by lmFit to report which coefficients in a linear model cannot be estimated.

Value

is.fullrank returns TRUE or FALSE.
nonEstimable returns a character vector of names for the columns of x which are linearly dependent on previous columns. If x has full column rank, then the value is NULL.

Author(s)

Gordon Smyth

Examples


# TRUE
is.fullrank(1)
is.fullrank(cbind(1,0:1))

# FALSE
is.fullrank(0)
is.fullrank(matrix(1,2,2))
nonEstimable(matrix(1,2,2))

[Package limma version 2.14.7 Index]