logsumexp {limma}	R Documentation

Log Sum of Exponentials

Description

Compute log( exp(x)+exp(y) ) without floating overflow or underflow

Usage

logsumexp(x, y)

Arguments

x
a numeric vector or matrix.

y
a numeric vector or matrix of same size as x.

Details

The computation uses logcosh().

Value

Numeric vector or matrix of same dimensions as x.

Author(s)

Gordon K Smyth

See Also

logcosh

Examples

x <- y <- c(1e-8,1e-7,1e-6,1e-5,1e-4,1,3,50,800)
logsumexp(x,y)
log( exp(x)+exp(y) )

[Package limma version 3.66.0 Index]