ma3x3 {limma}	R Documentation

Two dimensional Moving Averages with 3x3 Window

Description

Apply a specified function to each to each value of a matrix and its immediate neighbors.

Usage


ma3x3.matrix(x,FUN=mean,na.rm=TRUE,...)
ma3x3.spottedarray(x,printer,FUN=mean,na.rm=TRUE,...)

Arguments

x 	 numeric matrix

FUN 	 function to apply to each window of values

na.rm 	 logical value, should missing values be removed when applying FUN

... 	 other arguments are passed to FUN

printer 	 list giving the printer layout, see PrintLayout-class

Details

For ma3x3.matrix, x is an arbitrary function. for ma3x3.spotted, each column of x is assumed to contain the expression values of a spotted array in standard order. The printer layout information is used to re-arrange the values of each column as a spatial matrix before applying ma3x3.matrix.

Value

Numeric matrix of same dimension as x containing smoothed values

Author(s)

Gordon Smyth

See Also

An overview of functions for background correction are given in 04.Background.

Examples


x <- matrix(c(2,5,3,1,6,3,10,12,4,6,4,8,2,1,9,0),4,4)
ma3x3.matrix(x,FUN="mean")
ma3x3.matrix(x,FUN="min")

[Package limma version 2.14.7 Index]