makeUnique {limma}	R Documentation

Make Values of Character Vector Unique

Description

Paste characters on to values of a character vector to make them unique.

Usage

makeUnique(x)

Arguments

x
object to be coerced to a character vector

Details

Repeat values of x are labelled with suffixes "1", "2" etc.

Value

A character vector of the same length as x

Author(s)

Gordon Smyth

See Also

makeUnique is called by merge.RGList. Compare with make.unique in the base package.

Examples

x <- c("a","a","b")
makeUnique(x)

[Package limma version 3.66.0 Index]