MAList-class {limma}	R Documentation

M-value, A-value Expression List - class

Description

A simple list-based class for storing M-values and A-values for a batch of spotted microarrays. MAList objects are usually created during normalization by the functions normalizeWithinArrays or MA.RG.

Slots/List Components

MAList objects can be created by new("MAList",MA) where MA is a list. This class contains no slots (other than .Data), but objects should contain the following components:

M:	 numeric matrix containing the M-values (log-2 expression ratios). Rows correspond to spots and columns to arrays.

A:	 numeric matrix containing the A-values (average log-2 expression values).

Optional components include:

weights:	 numeric matrix of same dimensions as M containing relative spot quality weights. Elements should be non-negative.

other:	 list containing other matrices, all of the same dimensions as M.

genes:	 data.frame containing probe information. Should have one row for each spot. May have any number of columns.

targets:	 data.frame containing information on the target RNA samples. Rows correspond to arrays. May have any number of columns. Usually includes columns Cy3 and Cy5 specifying which RNA was hybridized to each array.

printer:	 list containing information on the process used to print the spots on the arrays. See PrintLayout.

Valid MAList objects may contain other optional components, but all probe or array information should be contained in the above components.

Methods

This class inherits directly from class list so any operation appropriate for lists will work on objects of this class. In addition, MAList objects can be subsetted and combined. RGList objects will return dimensions and hence functions such as dim, nrow and ncol are defined. MALists also inherit a show method from the virtual class LargeDataObject, which means that RGLists will print in a compact way.

Other functions in LIMMA which operate on MAList objects include normalizeWithinArrays, normalizeBetweenArrays, normalizeForPrintorder, plotMA and plotPrintTipLoess.

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of all the classes defined by this package.

marrayNorm is the corresponding class in the marray package.

[Package limma version 3.50.3 Index]