MArrayLM-class {limma}	R Documentation

Microarray Linear Model Fit - class

Description

A list-based class for storing the results of fitting gene-wise linear models to a batch of microarrays. Objects are normally created by lmFit.

Slots/Components

MArrayLM objects do not contain any slots (apart from .Data) but they should contain the following list components:

coefficients:

matrix containing fitted coefficients or contrasts

stdev.unscaled:

matrix containing unscaled standard deviations of the coefficients or contrasts

sigma:

numeric vector containing residual standard deviations for each gene

df.residual:

numeric vector containing residual degrees of freedom for each gene

Objects may also contain the following optional components:

Amean:

numeric vector containing the average log-intensity for each probe over all the arrays in the original linear model fit. Note this vector does not change when a contrast is applied to the fit using contrasts.fit.

genes:

data.frame containing gene names and annotation

design:

design matrix of full column rank

contrasts:

matrix defining contrasts of coefficients for which results are desired

F:

numeric vector giving moderated F-statistics for testing all contrasts equal to zero

F.p.value:

numeric vector giving p-value corresponding to F.stat

s2.prior:

numeric value giving empirical Bayes estimated prior value for residual variances

df.prior:

numeric vector giving empirical Bayes estimated degrees of freedom associated with s2.prior for each gene

s2.post:

numeric vector giving posterior residual variances

t:

matrix containing empirical Bayes t-statistics

var.prior:

numeric vector giving empirical Bayes estimated prior variance for each true coefficient

cov.coefficients:

numeric matrix giving the unscaled covariance matrix of the estimable coefficients

pivot:

integer vector giving the order of coefficients in cov.coefficients. Is computed by the QR-decomposition of the design matrix.

If there are no weights and no missing values, then the MArrayLM objects returned by lmFit will also contain the QR-decomposition of the design matrix, and any other components returned by lm.fit.

Methods

RGList objects will return dimensions and hence functions such as dim, nrow and ncol are defined. MArrayLM objects inherit a show method from the virtual class LargeDataObject.

The functions ebayes and classifyTestsF accept MArrayLM objects as arguments.

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of all the classes defined by this package.

[Package limma version 2.14.7 Index]