mdplot {limma}	R Documentation

mdplot

Description

Creates a mean-difference plot.

Usage


mdplot(x, ...)

Arguments

x 	 numeric matrix with at least two columns

... 	 any other arguments are passed to plot

Details

Plots differences vs means for a set of bivariate values. This is useful to contrast expression values for two microarrays.

Note that an MA-plot plotMA is a type of mean-difference plot.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

References

Chambers, J. M., Cleveland, W. S., Kleiner, B., and Tukey, P. A. (1983). Graphical Methods of Data Analysis. Wadsworth (pp. 48-57).

Cleveland, W. S., (1993). Visualizing Data. Hobart Press.

Bland, J. M., and Altman, D. G. (1986). Statistical methods for assessing agreement between two methods of clinical measurement. Lancet i, 307-310.

See also http://www.statsci.org/micrarra/refs/maplots.html

See Also

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 2.14.7 Index]