modifyWeights {limma}	R Documentation

modifyWeights

Description

Modify weights matrix for given gene status values.

Usage


modifyWeights(weights=rep(1,length(status)), status, values, multipliers)

Arguments

weights 	 numeric matrix of relative weights, rows corresponding to genes and columns to arrays

status 	 character vector giving the control status of each spot on the array, of same length as the number of rows of weights

values 	 character vector giving subset of the unique values of status

multipliers 	 numeric vector of same length as values giving factor by which weights will be modified

Details

The function is usually used to temporarily modify the weights matrix during normalization of data. The function can be used for example to give zero weight to spike-in ratio control spots during normalization.

Value

Numeric matrix of same dimensions as weights with rows corresponding to values in status modified by the specified multipliers.

Author(s)

Gordon Smyth

See Also

An overview of normalization functions available in LIMMA is given in 05.Normalization.

Examples


w <- matrix(runif(6*3),6,3)
status <- c("Gene","Gene","Ratio_Control","Ratio_Control","Gene","Gene")
modifyWeights(w,status,values="Ratio_Control",multipliers=0)

[Package limma version 2.14.7 Index]