mrlm {limma}	R Documentation

Fit Linear Model to Microrray Data by Robust Regression

Description

Fit a linear model genewise to expression data from a series of arrays. The fit is by robust M-estimation allowing for a small proportion of outliers. This is a utility function for lmFit.

Usage


mrlm(M,design=NULL,ndups=1,spacing=1,weights=NULL,...)

Arguments

M 	 numeric matrix containing log-ratio or log-expression values for a series of microarrays, rows correspond to genes and columns to arrays.

design 	 numeric design matrix defining the linear model, with rows corresponding to arrays and columns to comparisons to be estimated. The number of rows must match the number of columns of M. Defaults to the unit vector meaning that the arrays are treated as replicates.

ndups 	 a positive integer giving the number of times each gene is printed on an array. nrow(M) must be divisible by ndups.

spacing 	 the spacing between the rows of M corresponding to duplicate spots, spacing=1 for consecutive spots.

weights 	 numeric matrix of the same dimension as M containing weights. If it is of different dimension to M, it will be filled out to the same size. NULL is equivalent to equal weights.

... 	 any other arguments are passed to rlm.default.

Details

This is a utility function used by the higher level function lmFit. Most users should not use this function directly but should use lmFit instead.

This function fits a linear model for each gene by calling the function rlm from the MASS library.

Warning: don't use weights with this function unless you understand how rlm treats weights. The treatment of weights is somewhat different from that of lm.series and gls.series.

Value

A list with components

coefficients 	 numeric matrix containing the estimated coefficients for each linear model. Same number of rows as M, same number of columns as design.

stdev.unscaled 	 numeric matrix conformal with coef containing the unscaled standard deviations for the coefficient estimators. The standard errors are given by stdev.unscaled * sigma.

sigma 	 numeric vector containing the residual standard deviation for each gene.

df.residual 	 numeric vector giving the degrees of freedom corresponding to sigma.

qr 	 QR decomposition of design.

Author(s)

Gordon Smyth

See Also

rlm.

An overview of linear model functions in limma is given by 06.LinearModels.

[Package limma version 2.14.7 Index]