normalizeRobustSpline {limma}	R Documentation

Normalize Single Microarray Using Shrunk Robust Splines

Description

Normalize the M-values for a single microarray using robustly fitted regression splines and empirical Bayes shrinkage.

Usage


normalizeRobustSpline(M,A,layout,df=5,method="M")

Arguments

M 	 numeric vector of M-values

A 	 numeric vector of A-values

layout 	 list specifying the dimensions of the spot matrix and the grid matrix

df 	 degrees of freedom for regression spline, i.e., the number of regression coefficients and the number of knots

method 	 choices are "M" for M-estimation or "MM" for high breakdown point regression

Details

This function implements an idea similar to print-tip loess normalization but uses regression splines in place of the loess curves and uses empirical Bayes ideas to shrink the individual prtin-tip curves towards a common value. This allows the technique to introduce less noise into good quality arrays with little spatial variation while still giving good results on arrays with strong spatial variation.

Value

Numeric vector containing normalized M-values.

Author(s)

Gordon Smyth

References

The function is based on unpublished work by the author.

See Also

An overview of LIMMA functions for normalization is given in 05.Normalization.

Examples


library(sma)
data(MouseArray)
MA <- MA.RG(mouse.data)
normM <- normalizeRobustSpline(MA$M[,1],MA$A[,1],mouse.setup)

[Package limma version 2.14.7 Index]