normexp.signal {limma}	R Documentation

Expected Signal Given Observed Foreground Under Normal+Exp Model

Description

Adjust foreground intensities for observed background using Normal+Exp Model. This function is called by backgroundCorrect and is not normally called directly by the user.

Usage


normexp.signal(par, x)

Arguments

par 	 numeric vector containing the parameters of the Normal+Exp distribution, see normexp.fit

x 	 numeric vector of (background corrected) intensities

Details

In general the vector normmean is computed conditional on background at each spot.

Value

Numeric vector containing adjusted intensities.

Author(s)

Gordon Smyth

References

McGee, M., and Chen, Z. (2006). Parameter estimation for the exponential-normal convolution model for background correction of Affymetrix GeneChip data. Stat Appl Genet Mol Biol, 5(1), Article 24.

Ritchie, M. E., Silver, J., Oshlack, A., Silver, J., Holmes, M., Diyagama, D., Holloway, A., and Smyth, G. K. (2007). A comparison of background correction methods for two-colour microarrays. Bioinformatics http://bioinformatics.oxfordjournals.org/cgi/content/abstract/btm412

See Also

normexp.fit

An overview of background correction functions is given in 04.Background.

Examples

# See normexp.fit

[Package limma version 2.14.7 Index]