plotExons {limma}	R Documentation

Plot exons of differentially expressed gene

Description

Plot exons of differentially expressed gene and mark the differentially expressed exons.

Usage

plotExons(fit, coef = ncol(fit), geneid = NULL, genecolname = "GeneID",
          exoncolname = NULL, rank = 1L, FDR = 0.05)

Arguments

fit
MArrayLM fit object produced by eBayes.

coef
the coefficient (column) of fit for which differential expression is assessed.

geneid
character string, ID of the gene to plot.

genecolname
character string for the column name of fit$genes containing gene IDs. Defaults to "GeneID" for Entrez Gene ID.

exoncolname
character string for the column name of fit$genes containing exon IDs.

rank
integer, if geneid=NULL then this ranked gene will be plotted.

FDR
numeric, mark differentially expressed exons with false discovery rate less than this cutoff.

Details

Plots log2-fold-change by exon for the specified gene and highlight the differentially expressed exons. Show annotations such as GeneID, Symbol and Strand if available as title for the gene to plot. The significantly differentially expressed individual exons are highlighted as red dots for up-regulation and as blue dots for down-regulation. The size of the dots are weighted by its significance.

Value

A plot is created on the current graphics device.

Author(s)

Yifang Hu and Gordon Smyth

See Also

lmFit, eBayes, plotSplice

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples

## Not run: 
fit <- lmFit(y,design)
fit <- eBayes(fit)
plotExons(fit)
plotExons(fit, exoncolname = "Start", rank = 1)
plotExons(fit, geneid = "ps", genecolname = "Symbol", exoncolname = "Start")

## End(Not run)

[Package limma version 3.66.0 Index]