plotFB {limma}	R Documentation

FB-Plot

Description

Creates foreground-background plots.

Usage

## S3 method for class 'RGList'
plotFB(x, array = 1, lim = "separate", pch = 16, cex = 0.3,
     xlab = "log2 Background", ylab = "log2 Foreground", main = colnames(x)[array], ...)
## S3 method for class 'EListRaw'
plotFB(x, array = 1, pch = 16, cex=0.3,
     xlab = "log2 Background", ylab = "log2 Foreground", main = colnames(x)[array], ...)

Arguments

x
an RGList or EListRaw object.

array
integer giving the array to be plotted.

lim
character string indicating whether the red and green plots should have "separate" or "common" x- and y- co-ordinate limits.

pch
vector or list of plotting characters. Defaults to integer code 16.

cex
numeric vector of plot symbol expansions.

xlab
character string, label for x-axis.

ylab
character string, label for y-axis.

main
character string, title for plot.

...
any other arguments are passed to plot.

Details

A foreground-background plot is a plot of log2-foreground vs log2-background for a particular array. For two-color arrays, this function produces a pair of plots, one for the green channel and one for the red.

See points for possible values for pch, col and cex.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

See Also

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 3.66.0 Index]