plotFB {limma}	R Documentation

FB-Plot

Description

Creates foreground-background plots.

Usage


plotFB(RG, array=1, lim="separate", pch=16, cex=0.2, ...)

Arguments

RG 	 an RGList object.

array 	 integer giving the array to be plotted. Corresponds to columns of R, G, Rb and Gb.

lim 	 character string indicating whether the red and green plots should have "separate" or "common" x- and y- co-ordinate limits.

pch 	 vector or list of plotting characters. Defaults to integer code 16.

cex 	 numeric vector of plot symbol expansions.

... 	 any other arguments are passed to plot

Details

A foreground-background plot is a plot of log2-foreground vs log2-background for a particular channel on a particular two-color array. This function produces a pair of plots, one for green and one for red, for a specified array.

See points for possible values for pch, col and cex.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

See Also

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 2.14.7 Index]