plotSA {limma}	R Documentation

Sigma vs A plot for microarray linear model

Description

Plot residual standard deviation versus average log expression for a fitted microarray linear model.

Usage

plotSA(fit, xlab = "Average log-expression", ylab = "sqrt(sigma)", zero.weights = FALSE,
       pch = 16, cex = 0.3, col = c("black","red"), ...)

Arguments

fit
an MArrayLM object.

xlab
label for x-axis

ylab
label for y-axis

zero.weights
logical, should genes with all zero weights be plotted?

pch
vector of codes for plotting characters.

cex
numeric, vector of expansion factors for plotting characters.

col
plotting colors for regular and outlier variances respectively.

...
any other arguments are passed to plot

Details

This plot is used to check the mean-variance relationship of the expression data, after fitting a linear model. A scatterplot of residual-variances vs average log-expression is created. The plot is especially useful for examining the mean-variance trend estimated by eBayes or treat with trend=TRUE. It can be considered as a routine diagnostic plot in the limma-trend pipeline.

If robust empirical Bayes was used to create fit, then outlier variances are highlighted in the color given by col[2].

The y-axis is square-root fit$sigma, where sigma is the estimated residual standard deviation. The y-axis therefore corresponds to quarter-root variances. The y-axis was changed from log2-variance to quarter-root variance in limma version 3.31.21. The quarter-root scale matches the similar plot produced by the voom function and gives a better plot when some of the variances are close to zero.

See points for possible values for pch and cex.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

See Also

eBayes

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 3.66.0 Index]