plotSplice {limma}	R Documentation

Differential usage plot

Description

Plot relative log-fold changes by exons (or transcripts) for the specified gene and highlight differentially used exons (or transcripts).

Usage

plotSplice(fit, coef = ncol(fit), geneid = NULL, genecolname = NULL,
           rank = 1L, FDR = 0.05, xlab = "Exon")

Arguments

fit
MArrayLM fit object produced by diffSplice.

coef
the coefficient (column) of fit for which differentially splicing is assessed.

geneid
character string, ID of the gene to plot.

genecolname
column name of fit$genes containing gene IDs. Defaults to fit$genecolname.

rank
integer, if geneid=NULL then this ranked gene will be plotted.

FDR
numeric, highlight exons as red dots with false discovery rate less than this cutoff. The FDR of the individual exon is calculated based on the exon-level t-statistics test for differences between each exon and all other exons for the same gene.

xlab
x-axis lab for the plot. Typical choices are "Exon" for DEU analyses or "Transcript" for DTU analyses.

Details

Plot relative log2-fold-changes by exon (or transcript) for the specified gene. The relative logFC is the difference between that exon (or transcript)'s logFC and the overall logFC for the gene, as computed by diffSplice. Individual exons (or transcripts) that are significantly differentially used are highlighted as red dots. The size of the red dots increases with the statistical significance of the exon (or transcript).

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth and Yifang Hu

References

Baldoni PL, Chen L, Li M, Chen Y, Smyth GK (2025). Dividing out quantification uncertainty enables assessment of differential transcript usage with limma and edgeR. bioRxiv doi:10.1101/2025.04.07.647659.

Ritchie ME, Phipson B, Wu D, Hu Y, Law CW, Shi W, Smyth GK (2015). limma powers differential expression analyses for RNA-sequencing and microarray studies. Nucleic Acids Research 43, e47. doi:10.1093/nar/gkv007

See Also

diffSplice, topSplice

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples

# See diffSplice

[Package limma version 3.66.0 Index]