plotlines {limma}	R Documentation

plotlines

Description

Time course style plot of expression data.

Usage

plotlines(x,first.column.origin=FALSE,xlab="Column",ylab="x",col="black",lwd=1,...)

Arguments

x
numeric matrix or object containing expression data.

first.column.origin
logical, should the lines be started from zero?

xlab
x-axis label

ylab
y-axis label

col
vector of colors for lines

lwd
line width multiplier

...
any other arguments are passed to plot

Details

Plots a line for each probe.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

See Also

An overview of modeling functions and associated plots available in LIMMA is given in 06.LinearModels.

[Package limma version 3.66.0 Index]