plotMA3by2 {limma}	R Documentation

Write MA-Plots to Files

Description

Write MA-plots to files in PNG format, six plots to a file in a 3 by 2 grid arrangement.

Usage


plotMA3by2(MA, prefix="MA", path=NULL, main=colnames(MA), zero.weights=FALSE, common.lim=TRUE, device="png", ...)

Arguments

MA 	 an MAList or RGList object, or any list with components M containing log-ratios and A containing average intensities

prefix 	 character string giving prefix to attach to file names

path 	 character string specifying directory for output files

main 	 character vector giving titles for plots

zero.weights 	 logical, should points with non-positive weights be plotted

common.lim 	 logical, should all plots on a page use the same axis limits

device 	 device driver for the plot. Choices are "png", "jpeg", "pdf", "postscript".

... 	 any other arguments are passed to plotMA

Details

This function writes a series of graphic files to disk. Each file contains six MA-plots in three rows and two columns. The layout is optimized for A4-sized paper.

The graph format can be "png" or "jpeg", which are screen-resolution formats, or "pdf" or "postscript", which are loss-less formats. "png" is not available on every R platform. Note that "pdf" or "postscript" may produce very large files.

Value

No value is returned, but one or more files are written to the working directory. The number of files is determined by the number of columns of MA.

Author(s)

Gordon Smyth

See Also

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 2.14.7 Index]