plotPrintTipLoess {limma}	R Documentation

MA Plots by Print-Tip Group

Description

Creates a coplot giving MA-plots with loess curves by print-tip groups.

Usage

plotPrintTipLoess(object,layout,array=1,span=0.4,...)

Arguments

object
MAList or RGList object or list with components M containing log-ratios and A containing average intensities

layout
a list specifying the number of tip rows and columns and the number of spot rows and columns printed by each tip. Defaults to MA$printer if that is non-null.

array
integer giving the array to be plotted. Corresponds to columns of M and A.

span
span of window for lowess curve

...
other arguments passed to panel.smooth

Details

Note that spot quality weights in object are not used for computing the loess curves for this plot even though such weights would be used for loess normalization using normalizeWithinArrays.

Value

A plot is created on the current graphics device. If there are missing values in the data, then the vector of row numbers for spots with missing values is invisibly returned, as for coplot.

Author(s)

Gordon Smyth

See Also

An overview of diagnostic functions available in LIMMA is given in 09.Diagnostics.

[Package limma version 3.66.0 Index]