propexpr {limma}	R Documentation

Estimate Proportion of Expressed Probes

Description

Estimate the proportion of microarray probes which are expressed in each array.

Usage

propexpr(x, neg.x=NULL, status=x$genes$Status, labels=c("negative","regular"))

Arguments

x
matrix or similar object containing raw intensities for a set of arrays.

neg.x
matrix or similar object containing raw intensities for negative control probes for the same arrays. If NULL, then negative controls must be provided in x.

status
character vector specifying control type of each probe. Only used if neg.x is NULL.

labels
character vector giving the status values for negative control probes and regular (non-control) probes respectively. If of length 1, then all probes other than the negative controls are assumed to be regular. Only used if neg.x is NULL.

Details

This function estimates the overall proportion of probes on each microarray that are correspond to expressed genes using the method of Shi et al (2010). The function is especially useful for Illumina BeadChips arrays, although it can in principle be applied to any platform with good quality negative controls.

The negative controls can be supplied either as rows of x or as a separate matrix. If supplied as rows of x, then the negative controls are identified by the status vector. x might also include other types of control probes, but these will be ignored in the calculation.

Illumina BeadChip arrays contain 750~1600 negative control probes. If read.idat is used to read Illumina expression IDAT files, then the control probes will be populated as rows of the output EListRaw object, and the vector x$genes$Status will be set to identify control probes.

Alternatively, expression values can be exported from Illumina's GenomeStudio software as tab-delimited text files. In this case, the control probes are usually written to a separate file from the regular probes.

Value

Numeric vector giving the proportions of expressed probes in each array.

Author(s)

Wei Shi and Gordon Smyth

References

Shi W, de Graaf C, Kinkel S, Achtman A, Baldwin T, Schofield L, Scott H, Hilton D, Smyth GK (2010). Estimating the proportion of microarray probes expressed in an RNA sample. Nucleic Acids Research 38(7), 2168-2176. doi:10.1093/nar/gkp1204

See Also

Description to the control probes in Illumina BeadChips can be found in read.ilmn.

Examples

## Not run: 
# Read Illumina binary IDAT files
x <- read.idat(idat, bgx)
propexpr(x)

# Read text files exported from GenomeStudio
x <- read.ilmn(files = "sample probe profile.txt",
               ctrlfiles = "control probe profile.txt")
propexpr(x)

## End(Not run)

[Package limma version 3.66.0 Index]