read.ilmn {limma}	R Documentation

Read Illumina Expression Data

Description

Read Illumina summary probe profile files and summary control probe profile files

Usage

read.ilmn(files=NULL, ctrlfiles=NULL, path=NULL, ctrlpath=NULL, probeid="Probe",
          annotation=c("TargetID", "SYMBOL"), expr="AVG_Signal",
          other.columns="Detection", sep="\t", quote="\"", verbose=TRUE, ...)

Arguments

files
character vector giving the names of the summary probe profile files.

ctrlfiles
character vector giving the names of the summary control probe profile files.

path
character string giving the directory containing the summary probe profile files. Default is the current working directory.

ctrlpath
character string giving the directory containing the summary control probe profile files. Default is the same directory as for the probe profile files.

probeid
character string giving the name of the probe identifier column.

annotation
character vector giving possible column names for probe annotation.

expr
character string giving a keyword identifying the expression intensity columns. Any input column with column name containing this key will be read as containing intensity values.

other.columns
character vector giving keywords sufficient to identify any extra data columns that should be read in, such as "Detection", "Avg_NBEADS", "BEAD_STDEV" etc. The default of Detection is usually sufficient to identify the columns containing detection p-values.

sep
the field separator character.

quote
character string of characters to be treated as quote marks.

verbose
logical, TRUE to report names of profile files being read.

...
any other parameters are passed on to read.columns.

Details

Illumina BeadStudio ouputs probe intensities (regular probe intensities) and control probe intensities to summary probe profile files (containing regular probes) and summary control probe profile files, respectively. If both files and ctrlfiles are not NULL, this function will combine the data read from the two file types and save them to an EListRaw-class object. If one of them is NULL, then only the required data are read in.

Probe types are indicated in the Status column of genes, a component of the returned EListRaw-class object. There are totally seven types of control probes including negative, biotin, labeling, cy3_hyb, housekeeping, high_stringency_hyb or low_stringency_hyb. Regular probes have the probe type regular. The Status column will not be created if ctrlfiles is NULL.

To read in columns other than probeid, annotation and expr, users needs to specify keywords in other.columns. One keyword corresponds to one type of columns. Examples of keywords are "Detection", "Avg_NBEADS", "BEAD_STDEV" etc.

Value

An EListRaw-class object with the following components:

E
numeric matrix of intensities.

genes
data.frame of probe annotation. Contains any columns specified by annotation that are found in the input files.

other
a list of matrices corresponding to any other.columns found in the input files.

Author(s)

Wei Shi and Gordon K Smyth

See Also

read.ilmn.targets reads in Illumina expression data using the file information extracted from a target data frame which is often created by the readTargets function.

neqc performs normexp by control background correction, log transformation and quantile between-array normalization for Illumina expression data.

normexp.fit.control estimates the parameters of the normal+exponential convolution model with the help of negative control probes.

propexpr estimates the proportion of expressed probes in a microarray.

Examples

## Not run: 
x <- read.ilmn(files="sample probe profile.txt",
               ctrlfiles="control probe profile.txt")

## End(Not run)
# See neqc and beadCountWeights for other examples using read.ilmn

[Package limma version 3.66.0 Index]