read.maimages {limma}	R Documentation

Read RGList from Image Analysis Output Files

Description

Reads an RGList from a series of two-color microarray image analysis output files

Usage


read.maimages(files=NULL,source="generic",path=NULL,ext=NULL,names=NULL,columns=NULL,other.columns=NULL,annotation=NULL,wt.fun=NULL,verbose=TRUE,sep="\t",quote=NULL,...)
read.imagene(files,path=NULL,ext=NULL,names=NULL,columns=NULL,other.columns=NULL,wt.fun=NULL,verbose=TRUE,sep="\t",quote="\"",...)

Arguments

files 	 character vector giving the names of the files containing image analysis output or, for Imagene data, a character matrix of names of files. If omitted, then all files with extension ext in the specified directory will be read in alphabetical order.

source 	 character string specifying the image analysis program which produced the output files. Choices are "generic", "agilent", "arrayvision", arrayvision.ARM", arrayvision.MTM", "bluefuse", "genepix", "genepix.custom", "genepix.median", "imagene", "quantarray", "scanarrayexpress", "smd.old", "smd", "spot" or "spot.close.open".

path 	 character string giving the directory containing the files. The default is the current working directory.

ext 	 character string giving optional extension to be added to each file name

names 	 character vector of names to be associated with each array as column name. Defaults to removeExt(files).

columns 	 list with fields R, G, Rb and Gb giving the column names to be used for red and green foreground and background or, in the case of Imagene data, a list with fields f and b. This argument is optional if source is specified, otherwise it is required.

other.columns 	 character vector of names of other columns to be read containing spot-specific information

annotation 	 character vector of names of columns containing annotation information about the probes

wt.fun 	 function to calculate spot quality weights

verbose 	 logical, TRUE to report each time a file is read

sep 	 the field separator character

quote 	 character string of characters to be treated as quote marks

... 	 any other arguments are passed to read.table

Details

This is the main data input function for the LIMMA package for two-color microarray data. It extracts the foreground and background intensities from a series of files, produced by an image analysis program, and assembles them into the components of one list. The image analysis programs Agilent Feature Extraction, ArrayVision, BlueFuse, GenePix, ImaGene, QuantArray (Version 3 or later), Stanford Microarray Database (SMD) and SPOT are supported explicitly. Data from some other image analysis programs can be read if the appropriate column names containing the foreground and background intensities are specified using the columns argument. (This will work if the column names are unique and if there are no rows in the file after the last line of data. Header lines are ok.)

SMD data should consist of raw data files from the database, in tab-delimited text form. There are two possible sets of column names depending on whether the data was entered into the database before or after September 2003. source="smd.old" indicates that column headings in use prior to September 2003 should be used. In the case of GenePix, two possible foreground estimators are supported: source="genepix" uses the mean foreground estimates while source="genepix.median" uses median foreground estimates. GenePix 6.0 and later also supplies some custom background options, notably morphological background. If the GPR files have been written using a custom background, you may read it using source="genepix.custom". In the case of SPOT, two possible background estimators are supported: if source="spot.close.open" then background intensities are estimated from morph.close.open rather than morph.

Spot quality weights may be extracted from the image analysis files using a weight function wt.fun. wt.fun may be any user-supplied function which accepts a data.frame argument and returns a vector of non-negative weights. The columns of the data.frame are as in the image analysis output files. There is one restriction, which is that the column names should be refered to in full form in the weight function, i.e., do not rely on name expansion for partial matches when refering to the names of the columns. See QualityWeights for suggested weight functions.

For Imagene image data the argument files should be a matrix with two columns. The first column should contain the names of the files containing green channel (cy3) data and the second column should contain names of files containing red channel (cy5) data. If source="imagene" and files is a vector of even length instead of a matrix, then each consecutive pair of file names is assumed to correspond to the same array. The function read.imagene is called by read.maimages when source="imagene". It does not need to be called directly by users.

ArrayVision reports spot intensities in a number of different ways. read.maimages caters for ArrayVision's Artifact-removed (ARM) density values as "arrayvision.ARM" or for Median-based Trimmed Mean (MTM) density values as "arrayvision.MTM". ArrayVision users may find it useful to read the top two lines of their data file to check which version of density values they have.

The argument other.columns allows arbitrary columns of the image analysis output files to be preserved in the data object. These become matrices in the component other component. For ImaGene data, the other column headings with be prefixed with "R " or "G " as appropriate.

Value

An RGList object containing the components

R 	 matrix containing the red channel foreground intensities for each spot for each array.

Rb 	 matrix containing the red channel background intensities for each spot for each array.

G 	 matrix containing the green channel foreground intensities for each spot for each array.

Gb 	 matrix containing the green channel background intensities for each spot for each array.

weights 	 spot quality weights, if wt.fun is given

other 	 list containing matrices corresponding to other.columns if given

genes 	 data frame containing annotation information about the probes, for example gene names and IDs and spatial positions on the array, currently set only if source is "agilent", "genepix" or source="imagene" or if the annotation argument is set

targets 	 data frame with column FileName giving the names of the files read

source 	 character string giving the image analysis program name

printer 	 list of class PrintLayout, currently set only if source="imagene"

Warnings

All image analysis files being read are assumed to contain data for the same genelist in the same order. No checking is done to confirm that this is true. Probe annotation information is read from the first file only.

Author(s)

Gordon Smyth, with speed improvements by Marcus Davy

References

Web pages for the image analysis software packages mentioned here are listed at http://www.statsci.org/micrarra/image.html

See Also

read.maimages is based on read.table in the base package. read.marrayRaw is the corresponding function in the marray package.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


#  Read all .gpr files from current working directory
#  and give weight 0.1 to spots with negative flags

## Not run: 
files <- dir(pattern="*\\.gpr$")
RG <- read.maimages(files,"genepix",wt.fun=wtflags(0.1))
## End(Not run)

#  Read all .spot files from current working director and down-weight
#  spots smaller or larger than 150 pixels

## Not run: 
files <- dir(pattern="*\\.spot$")
RG <- read.maimages(files,"spot",wt.fun=wtarea(150))
## End(Not run)

[Package limma version 2.14.7 Index]