readHeader {limma}	R Documentation

Read Header Information from Image Analysis Raw Data File

Description

Read the header information from a GenePix Results (GPR) file or from an SMD raw data file. These functions are used internally by read.maimages and are not usually called directly by users.

Usage


readGenericHeader(file, columns, sep="\t")
readGPRHeader(file)
readSMDHeader(file)

Arguments

file 	 character string giving file name. If it does not contain an absolute path, the file name is relative to the current working directory.

columns 	 character vector specifying data column headings expected to be in file

sep 	 the character string separating column names

Details

Raw data files exported by image analysis programs include a number of header lines which contain information about the scanning process. This function extracts that information and locates the line where the intensity data begins. readGPRHeader is for GenePix output and readSMDHeader is for files from the Stanford Microarray Database (SMD). readGenericHeader finds the line in the file on which the data begins by searching for specified column headings.

Value

A list with components corresponds to lines of header information. A key component is NHeaderRecords which gives the number of lines in the file before the intensity data begins. All other components are character vectors.

Author(s)

Gordon Smyth

References

See http://www.axon.com/gn_GenePix_File_Formats.html for GenePix formats.

See http://www.bluegnome.co.uk for information on BlueFuse.

See http://genome-www.stanford.edu/Microarray for the SMD.

See Also

read.maimages

An overview of LIMMA functions to read data is given in 03.ReadingData.

[Package limma version 2.14.7 Index]