readSpotTypes {limma}	R Documentation

Read Spot Types File

Description

Read a table giving regular expressions to identify different types of spots in the gene-dataframe.

Usage

readSpotTypes(file="SpotTypes.txt",path=NULL,sep="\t",check.names=FALSE,...)

Arguments

file
character string giving the name of the file specifying the spot types.

path
character string giving the directory containing the file. Can be omitted if the file is in the current working irectory.

sep
the field separator character

check.names
logical, if FALSE column names will not be converted to valid variable names, for example spaces in column names will not be left as is

...
any other arguments are passed to read.table

Details

The file is a text file with rows corresponding to types of spots and the following columns: SpotType gives the name for the spot type, ID is a regular expression matching the ID column, Name is a regular expression matching the Name column, and Color is the R name for the color to be associated with this type.

Value

A data frame with columns

SpotType
character vector giving names of the spot types

ID
character vector giving regular expressions

Name
character vector giving regular expressions

Color
character vector giving names of colors

Author(s)

Gordon Smyth following idea of James Wettenhall

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

[Package limma version 3.66.0 Index]