removeExt {limma}	R Documentation

Remove Common Extension from File Names

Description

Finds and removes any common extension from a vector of file names.

Usage


removeExt(x)

Arguments

x 	 character vector

Value

A character vector of the same length as x in which any common extension has been stripped off.

Author(s)

Gordon Smyth

See Also

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


x <- c("slide1.spot","slide2.spot","slide3.spot")
removeExt(x)

[Package limma version 2.14.7 Index]