RGList-class {limma}	R Documentation

Red, Green Intensity List - class

Description

A simple list-based class for storing red and green channel foreground and background intensities for a batch of spotted microarrays. RGList objects are normally created by read.maimages.

Slots/List Components

RGList objects can be created by new("RGList",RG) where RG is a list. Objects of this class contains no slots (other than .Data), but objects should contain the following list components:

R: 	 numeric matrix containing the red (cy5) foreground intensities. Rows correspond to spots and columns to arrays.

G: 	 numeric matrix containing the green (cy3) foreground intensities. Rows correspond to spots and columns to arrays.

Optional components include

Rb: 	 numeric matrix containing the red (cy5) background intensities

Gb: 	 numeric matrix containing the green (cy3) background intensities

weights: 	 numeric matrix of same dimension as R containing relative spot quality weights. Elements should be non-negative.

other: 	 list containing numeric matrices of other spot-specific information.

genes: 	 data.frame containing probe information. Should have one row for each spot. May have any number of columns.

targets: 	 data.frame containing information on the target RNA samples. Rows correspond to arrays. May have any number of columns. Usually includes columns Cy3 and Cy5 specifying which RNA was hybridized to each array.

printer: 	 list containing information on the process used to print the spots on the arrays. See PrintLayout.

Valid RGList objects may contain other optional components, but all probe or array information should be contained in the above components.

Methods

This class inherits directly from class list so any operation appropriate for lists will work on objects of this class. In addition, RGList objects can be subsetted, combined and merged. RGList objects will return dimensions and hence functions such as dim, nrow and ncol are defined. RGLists also inherit a show method from the virtual class LargeDataObject, which means that RGLists will print in a compact way.

RGList objects can be converted to exprSet2 objects by as(RG,"exprSet2").

Other functions in LIMMA which operate on RGList objects include normalizeBetweenArrays, normalizeForPrintorder, normalizeWithinArrays.

Author(s)

Gordon Smyth

See Also

02.Classes gives an overview of all the classes defined by this package.

marrayRaw-class is the corresponding class in the marrayClasses package.

[Package limma version 2.14.7 Index]