squeezeVar {limma}	R Documentation

Squeeze Sample Variances

Description

Squeeze a set of sample variances together by computing empirical Bayes posterior means.

Usage

squeezeVar(var, df, covariate = NULL, span = NULL,
           robust = FALSE, winsor.tail.p = c(0.05,0.1), legacy = NULL)

Arguments

var
numeric vector of independent sample variances.

df
numeric vector of degrees of freedom for the sample variances. Can be a unit vector or of same length as var.

covariate
numeric covariate of same length as var for estimating a trended prior variance. If NULL, then the prior variance var.prior is constant.

span
lowess span used for prior variance trend.

robust
logical, should the estimation of df.prior and var.prior be robustified against outlier sample variances?

winsor.tail.p
numeric vector of length 1 or 2, giving left and right tail proportions of x to Winsorize when robust=TRUE.

legacy
logical. If FALSE then the new function fitFDistUnequalDF1 will be called internally, otherwise the legacy functions fitFDist or fitFDistRobustly will be used. If NULL, then fitFDistUnequalDF1 will be used whenever the degrees of freedom df are not all equal. If span is set, then this argument will be ignored and fitFDistUnequalDF1 will be used.

Details

This function implements empirical Bayes algorithms proposed by Smyth (2004) and Phipson et al (2016).

A conjugate Bayesian hierarchical model is assumed for a set of sample variances. The hyperparameters are estimated by fitting a scaled F-distribution to the sample variances. The function returns the posterior variances and the estimated hyperparameters.

Specifically, the sample variances var are assumed to follow scaled chi-squared distributions, conditional on the true variances, and an scaled inverse chi-squared prior is assumed for the true variances. The scale and degrees of freedom of this prior distribution are estimated from the values of var.

The effect of this function is to squeeze the variances towards a common value, or to a global trend if a covariate is provided. The squeezed variances have a smaller expected mean square error to the true variances than do the sample variances themselves.

The amount of squeezing is controlled by the prior.df. Both the global trend and the prior df are estimated internally but fitting an F-distribution to the sample variances, using either fitFDist() or fitFDistRobustly() or fitFDistUnequalDF1().

If covariate is non-null, then the scale parameter of the prior distribution is assumed to depend on the covariate. If the covariate is average log-expression, then the effect is an intensity-dependent trend similar to that in Sartor et al (2006).

robust=TRUE implements the robust empirical Bayes procedure of Phipson et al (2016), which allows some of the var values to be outliers.

The legacy argument was added in limma version 3.61.8 (August 2024). If legacy=FALSE, then the new function fitFDistUnequalDF1() provides improved estimation of the global trend and prior df hyperparameters, especially when the df values are unequal. legacy=TRUE provides legacy behavior for backward compatibility.

Value

A list with components

var.post
numeric vector of posterior variances. Of same length as var.

var.prior
location or scale of prior distribution. A vector of same length as var if covariate is non-NULL, otherwise a single value.

df.prior
degrees of freedom of prior distribution. A vector of same length as var if robust=TRUE, otherwise a single value.

Note

This function is called by eBayes, but beware a possible confusion with the output from that function. The values var.prior and var.post output by squeezeVar correspond to the quantities s2.prior and s2.post output by eBayes, whereas var.prior output by eBayes relates to a different parameter.

Author(s)

Gordon Smyth

References

Phipson B, Lee S, Majewski IJ, Alexander WS, and Smyth GK (2016). Robust hyperparameter estimation protects against hypervariable genes and improves power to detect differential expression. Annals of Applied Statistics 10, 946-963. doi:10.1214/16-AOAS920

Sartor MA, Tomlinson CR, Wesselkamper SC, Sivaganesan S, Leikauf GD, Medvedovic M (2006). Intensity-based hierarchical Bayes method improves testing for differentially expressed genes in microarray experiments. BMC bioinformatics 7, 538.

Smyth GK (2004). Linear models and empirical Bayes methods for assessing differential expression in microarray experiments. Statistical Applications in Genetics and Molecular Biology Volume 3, Issue 1, Article 3. doi:10.2202/1544-6115.1027. See also the Preprint Version https://gksmyth.github.io/pubs/ebayes.pdf incorporating corrections to 30 June 2009.

See Also

This function is called by eBayes.

This function calls fitFDist, fitFDistRobustly or fitFDistUnequalDF1 .

An overview of linear model functions in limma is given by 06.LinearModels.

Examples

s2 <- rchisq(20,df=5)/5
squeezeVar(s2, df=5)

[Package limma version 3.66.0 Index]