strsplit2 {limma}	R Documentation

Split Composite Names

Description

Split a vector of composite names into a matrix of simple names.

Usage


strsplit2(x, split, extended = TRUE, fixed = FALSE, perl = FALSE)

Arguments

x 	 character vector

split 	 character to split each element of vector on, see strsplit

extended 	 logical. If TRUE, extended regular expression matching is used, see strsplit.

fixed 	 logical. If TRUE match string exactly, otherwise use regular expressions. Has priority over perl and extended.

perl 	 logical. Should perl-compatible regexps be used? Has priority over extended.

Details

This function is the same as strsplit except that the output value is a matrix instead of a list. The first column of the matrix contains the first component from each element of x, the second column contains the second components etc. The number of columns is equal to the maximum number of components for any element of x.

Value

A list containing components

Name 	 character vector of the same length as x contain first splits of each element

Annotation 	 character vector of the same length as x contain second splits of each element

Author(s)

Gordon Smyth

See Also

strsplit.

An overview of LIMMA functions for reading data is given in 03.ReadingData.

Examples


x <- c("AA196000;actinin, alpha 3",
"AA464163;acyl-Coenzyme A dehydrogenase, very long chain",
"3E7;W15277;No Annotation")
strsplit2(x,split=";")

[Package limma version 2.14.7 Index]