subsetting {limma}	R Documentation

Subset RGList, MAList or MArrayLM Objects

Description

Extract a subset of an RGList, MAList or MArrayLM object.

Usage


## S3 method for class 'RGList':
object[i, j, ...]

Arguments

object 	 object of class RGList, MAList or MArrayLM

i,j 	 elements to extract. i subsets the genes or spots while j subsets the arrays

... 	 not used

Details

i,j may take any values acceptable for the matrix components of object. See the Extract help entry for more details on subsetting matrices.

Value

An object of the same class as object holding data from the specified subset of genes and arrays.

Author(s)

Gordon Smyth

See Also

Extract in the base package.

03.ReadingData gives an overview of data input and manipulation functions in LIMMA.

Examples


M <- A <- matrix(11:14,4,2)
rownames(M) <- rownames(A) <- c("a","b","c","d")
colnames(M) <- colnames(A) <- c("A","B")
MA <- new("MAList",list(M=M,A=A))
MA[1:2,]
MA[1:2,2]
MA[,2]

[Package limma version 2.14.7 Index]