summary {limma}	R Documentation

Summaries of Microarray Data Objects

Description

Briefly summarize microarray data objects.

Usage

## S3 method for class 'RGList'
summary(object, ...)

Arguments

object
an object of class RGList, MAList, EListRaw, EList or MArrayLM

...
other arguments are not used

Details

The data objects are summarized as if they were lists, i.e., brief information about the length and type of the components is given.

Value

A table.

Author(s)

Gordon Smyth

See Also

summary in the base package.

02.Classes gives an overview of data classes used in LIMMA.

[Package limma version 3.66.0 Index]