targetsA2C {limma}	R Documentation

Convert Two-Color Targets Dataframe from One-Row-Per-Array to One-Row-Per-Channel

Description

Convert a two-color targets dataframe with one row per array to one with one row per channel.

Usage

targetsA2C(targets, channel.codes = c(1,2), channel.columns = list(Target=c("Cy3","Cy5")),
           grep = FALSE)

Arguments

targets
data.frame with one row per array giving information about target samples associated covariates.

channel.codes
numeric or character vector of length 2 giving codes for the channels

channel.columns
named list of character vectors of length 2. Each entry gives a pair of names of columns in targets which contain channel-specific information. This pair of columns should be assembled into one column in the output.

grep
logical, if TRUE then the channel column names are found by greping, i.e., the actual column names need only contain the names given by channel.columns as substrings

Details

The targets dataframe holds information about the RNA samples used as targets in the microarray experiment. It is often read from a file using readTargets. This function is used to convert the dataframe from an array-orientated format with one row for each array and two columns for the two channels into a channel-orientated format with one row for each individual channel observations. In statistical terms, the first format treats the arrays as cases and treats the channels as repeated measurements. The second format treats the individual channel observations as cases. The second format may be more appropriate if the data is to be analyzed in terms of individual log-intensities.

Value

data.frame with twice as many rows as targets. Any pair of columns named by channel.columns will now be one column.

Author(s)

Gordon Smyth

References

Smyth, GK, and Altman, NS (2013). Separate-channel analysis of two-channel microarrays: recovering inter-spot information. BMC Bioinformatics 14, 165. doi:10.1186/1471-2105-14-165

See Also

targetsA2C is used by the coerce method from RGList to ExpressionSet in the convert package.

An overview of methods for single channel analysis in limma is given by 07.SingleChannel.

Examples

targets <- data.frame(FileName=c("file1.gpr","file2.gpr"),Cy3=c("WT","KO"),Cy5=c("KO","WT"))
targetsA2C(targets)

[Package limma version 3.66.0 Index]