tmixture {limma}	R Documentation

Estimate Scale Factor in Mixture of t-Distributions

Description

These functions estimate the unscaled standard deviation of the true (unobserved) log fold changes for differentially expressed genes. They are used internally by the eBayes function and are not intended to be called directly by users.

Usage

tmixture.vector(tstat, stdev.unscaled, df, proportion, v0.lim = NULL)
tmixture.matrix(tstat, stdev.unscaled, df, proportion, v0.lim = NULL)

Arguments

tstat
numeric vector or matrix of t-statistics. tmixture.vector assumes a vector while tmixture.matrix assumes a matrix.

stdev.unscaled
numeric vector or matrix, conformal with tstat, containing the unscaled standard deviations of the coefficients used to compute the t-statistics.

df
numeric vector giving the degrees of freedom associated with tstat.

proportion
assumed proportion of genes that are differentially expressed.

v0.lim
numeric vector of length 2 giving the lower and upper limits for the estimated unscaled standard deviations.

Details

The values in each column of tstat are assumed to follow a mixture of an ordinary t-distribution, with mixing proportion 1-proportion, and (v0+v1)/v1 times a t-distribution, with mixing proportion proportion. Here v1 is stdev.unscaled^2 and v0 is the value to be estimated.

Value

Numeric vector, of length equal to the number of columns of tstat, containing estimated v0 values.

Author(s)

Gordon Smyth

See Also

eBayes

[Package limma version 3.66.0 Index]