tmixture {limma}	R Documentation

Estimate Scale Factor in Mixture of t-Distributions

Description

This function estimates the unscaled standard deviation of the log fold change for differentially expressed genes. It is called by the function ebayes and is not intended to be called by users.

Usage


tmixture.vector(tstat,stdev.unscaled,df,proportion,v0.lim=NULL)
tmixture.matrix(tstat,stdev.unscaled,df,proportion,v0.lim=NULL)

Arguments

tstat 	 numeric vector or matrix of t-statistics

stdev.unscaled 	 numeric matrix conformal with tstatf containing the unscaled standard deviations for the coefficient estimators

df 	 numeric vector giving the degrees of freedom associated with tstat

proportion 	 assumed proportion of genes which are differentially expressed

v0.lim 	 numeric vector of length 2, assumed lower and upper limits for the estimated unscaled standard deviation

Details

The values in each column of tstat are assumed to follow a mixture of an ordinary t-distribution, with mixing proportion 1-proportion, and (v0+v1)/v1 times a t-distribution, with mixing proportion proportion. Here v1=stdev.unscaled^2 and v0 is the value to be estimated.

Value

Numeric vector of length equal to the number of columns of tstat and stdev.unscaled.

Author(s)

Gordon Smyth

See Also

ebayes

[Package limma version 2.14.7 Index]