topGO {limma}	R Documentation

Table of Top GO Terms or Top KEGG Pathways

Description

Extract top GO terms from goana output or top KEGG pathways from kegga output.

Usage

topGO(results, ontology = c("BP", "CC", "MF"), sort = NULL, number = 20L, 
      truncate.term = NULL, p.value = 1)
topKEGG(results, sort = NULL, number = 20L, truncate.path = NULL, p.value = 1)

Arguments

results
data frame produced by goana or kegga.

ontology
character vector of ontologies to be included in output. Elements should be one or more of "BP", "CC" or "MF".

sort
character vector of names of gene lists for which results are required. Should be one or more of the column names of results. Defaults to all gene lists.

number
maximum number of top GO terms or top KEGG pathways to list. For all terms or all pathways, set number=Inf.

truncate.term
truncate the name of the GO term at this number of characters.

truncate.path
truncate the name of the KEGG pathway at this number of characters.

p.value
p.value cutoff. Only GO terms or pathways with lower p-values are included in the output.

Details

topGO organizes the output from goana into top-tables of the most significant GO terms. topKEGG similarly extracts the most significant KEGG pathways from kegga output. In either case, rows are sorted by the minimum p-value of any of the result columns specified by sort.

Value

Same as results but with rows subsetted by Ontology and sorted by p-value.

Author(s)

Gordon Smyth and Yifang Hu

See Also

goana, kegga

See 10.GeneSetTests for a description of other functions used for gene set testing.

Examples

# See goana examples

[Package limma version 3.66.0 Index]