topRomer {limma}	R Documentation

Top Gene Set Testing Results from Romer

Description

Extract a matrix of the top gene set testing results from the romer output.

Usage

topRomer(x,n=10,alternative="up")

Arguments

x
matrix which is the output from romer.

n
number of top gene set testing results to be extracted.

alternative
character which can be one of the three possible alternative p values: "up", "down" or "mixed".

Details

This function takes the results from romer and returns a number of top gene set testing results that are sorted by the p values.

Value

matrix, which is sorted by the "up", "down" or "mixed" p values, with the rows corresponding to estimated p-values for the top number of gene sets and the columns corresponding to the number of genes for each gene set and the alternative hypotheses mixed, up, down.

Author(s)

Gordon Smyth and Yifang Hu

See Also

romer

There is a topic page on 10.GeneSetTests.

Examples

# See romer for examples

[Package limma version 3.66.0 Index]