topSplice {limma}	R Documentation

Top Table of Differential Transcript Usage

Description

Top table of results from a diffSplice differential transcript usage analysis.

Usage

topSplice(fit, coef = ncol(fit), test = "F", number = 10, FDR=1, sort.by = "p")

Arguments

fit
MArrayLM fit object produced by diffSplice.

coef
the coefficient (column) of fit for which differential usage is assessed.

test
character string specifying which statistical test to apply. Possible values are "simes", "F" or "t". "F" gives F-tests for each gene. "t" gives t-tests for each transcript or exon. "simes" gives genewise p-values derived from the t-tests after Simes adjustment for each gene.

number
integer, maximum number of rows to output.

FDR
numeric, only show exons or genes with false discovery rate less than this cutoff.

sort.by
character string specifying which column to sort results by. Possible values for "p", "logFC", "NExons" or "none". "logFC" is only available if test="t" and "NExons" is only available if test="simes" or test="F".

Details

This function ranks results from a differential transcript usage analysis conducted by diffSplice. The output object from diffSplice includes results at the gene level and results at the isoform level. The isoform-level results can be transcripts, if a DTU analysis was done, or can be exons and exon-junctions, if a DEU analysis was done.

Choosing test="F" or test="simes" gives gene-level results, where the null hypothesis is that all transcripts or exons are used equally. Choosing test="t" gives transcript or exon level results.

The F-statistic option chosen by test="F" conducts an ANOVA-like F-tests of differential usage for each gene. Simes method chosen by test="simes" processes the isoform-level p-values for each gene to give an overall call of differential usage for that gene. The Simes p-values is likely to be more powerful than F-tests when only one transcript of the gene is differentially expressed. The F-tests are likely to be powerful than Simes when there multiple transcripts of the same gene differentially expressed in different directions.

Value

A data.frame with annotation columns found in fit plus the following columns

logFC
log2-fold change of transcript vs other transcripts for the same gene (if test="t")

t
moderated t-statistic (if test="t")

F
moderated F-statistic (if test="F")

P.Value
p-value

FDR
false discovery rate bound, computed by the Benjamini & Hochberg method.

Author(s)

Gordon Smyth

References

Baldoni PL, Chen L, Li M, Chen Y, Smyth GK (2025). Dividing out quantification uncertainty enables assessment of differential transcript usage with limma and edgeR. bioRxiv doi:10.1101/2025.04.07.647659.

See Also

diffSplice, plotSplice. See also the diffSplice method for DGEGLM objects in the edgeR package.

A summary of functions available in LIMMA for RNA-seq analysis is given in 11.RNAseq.

Examples

# See diffSplice

[Package limma version 3.66.0 Index]