trigammaInverse {limma}	R Documentation

Inverse Trigamma Function

Description

The inverse of the trigamma function.

Usage


trigammaInverse(x)

Arguments

x 	 numeric vector or array

Details

The function uses Newton's method with a clever starting value to ensure monotonic convergence.

Value

Numeric vector or array y satisfying trigamma(y)==x.

Note

This function does not accept a data.frame as argument although the internal function trigamma does.

Author(s)

Gordon Smyth

See Also

trigamma

Examples


y <- trigammaInverse(5)
trigamma(y)

[Package limma version 2.14.7 Index]