uniquegenelist {limma}	R Documentation

Eliminate Duplicate Names from the Gene List

Description

Eliminate duplicate names from the gene list. The new list is shorter than the full list by a factor of ndups.

Usage


uniquegenelist(genelist,ndups=2,spacing=1)

Arguments

genelist 	 vector of gene names

ndups 	 number of duplicate spots. The number of rows of genelist must be divisible by ndups.

spacing 	 the spacing between duplicate names in genelist

Value

A vector of length length(genelist)/ndups containing each gene name once only.

Author(s)

Gordon Smyth

See Also

unwrapdups

Examples


genelist <- c("A","A","B","B","C","C","D","D")
uniquegenelist(genelist,ndups=2)
genelist <- c("A","B","A","B","C","D","C","D")
uniquegenelist(genelist,ndups=2,spacing=2)

[Package limma version 2.14.7 Index]