venn {limma}	R Documentation

Venn Diagrams

Description

Compute classification counts or plot classification counts in a Venn diagram.

Usage


vennCounts(x, include="both")
vennDiagram(object, include="both", names, mar=rep(1,4), cex=1.5, lwd=1,
circle.col, counts.col, show.include, ...)

Arguments

x 	 numeric matrix of 0's and 1's indicating significance of a test. Usually created by decideTests.

object 	 either a TestResults matrix or a VennCounts object produced by vennCounts.

include 	 character string specifying whether to counts genes up-regulated, down-regulated or both. See details. Choices are "both", "up" or "down".

names 	 optional character vector giving names for the sets or contrasts

mar 	 numeric vector of length 4 specifying the width of the margins around the plot. This argument is passed to par.

cex 	 numerical value giving the amount by which the contrast names should be scaled on the plot relative to the default.plotting text. See par.

lwd 	 numerical value giving the amount by which the circles should be scaled on the plot. See par.

circle.col 	 optional vector of color specifications defining the colors by which the circles should be drawn. See par.

counts.col 	 optional vector of color specifications defining the colors by which the circles should be drawn. See par.

show.include 	 logical value whether the value of include should be printed on the plot. Defaults to FALSE if include is a single value and TRUE otherwise

... 	 any other arguments are passed to plot

Details

If a vennCounts object is given to vennDiagram, the include parameter is ignored. If a TestResults object is given, then it is possible to set include as a vector of 2 character strings and both will be shown.

Value

vennCounts produces a VennCounts object, which is a numeric matrix with last column "Counts" giving counts for each possible vector outcome. vennDiagram causes a plot to be produced on the current graphical device. For venDiagram, the number of columns of object should be three or fewer.

Author(s)

Gordon Smyth James Wettenhall and Francois Pepin

See Also

An overview of linear model functions in limma is given by 06.LinearModels.

Examples


Y <- matrix(rnorm(100*6),100,6)
Y[1:10,3:4] <- Y[1:10,3:4]+3
Y[1:20,5:6] <- Y[1:20,5:6]+3
design <- cbind(1,c(0,0,1,1,0,0),c(0,0,0,0,1,1))
fit <- eBayes(lmFit(Y,design))
results <- decideTests(fit)
a <- vennCounts(results)
print(a)
vennDiagram(a)
vennDiagram(results,include=c("up","down"),counts.col=c("red","green"))

[Package limma version 2.14.7 Index]