volcanoplot {limma}	R Documentation

Volcano Plot

Description

Creates a volcano plot for a specified coefficient of a linear model.

Usage

volcanoplot(fit, coef = 1, style = "p-value",
            highlight = 0, names = fit$genes$ID, hl.col = "blue",
            xlab = "Log2 Fold Change", ylab = NULL, pch=16, cex=0.35, ...)

Arguments

fit
an MArrayLM fitted linear model object.

coef
index indicating which coefficient of the linear model is to be plotted.

style
character string indicating which significance statistic to plot on the y-axis. Possibilities are "p-value" or "B-statistic".

highlight
number of top genes to be highlighted by name.

names
character vector of length nrow(fit) giving gene names. Only used if highlight > 0.

hl.col
color for the gene names. Only used if highlight > 0.

xlab
character string giving label for x-axis

ylab
character string giving label for y-axis

pch
vector or list of plotting characters.

cex
numeric vector of plot symbol expansions.

...
any other arguments are passed to plot

Details

A volcano plot displays log fold changes on the x-axis versus a measure of statistical significance on the y-axis. Here the significance measure can be -log(p-value) or the B-statistics, which give the posterior log-odds of differential expression.

The plot is optionally annotated with the names of the most significant genes.

Value

No value is returned but a plot is created on the current graphics device.

Author(s)

Gordon Smyth

See Also

An overview of presentation plots following the fitting of a linear model in LIMMA is given in 06.LinearModels.

Examples

#  See lmFit examples

[Package limma version 3.66.0 Index]