volcanoplot {limma}	R Documentation

Volcano Plot

Description

Creates a volcano plot of log-fold changes versus log-odds of differential expression.

Usage


volcanoplot(fit, coef=1, highlight=0, names=fit$genes$ID, xlab="Log Fold Change", ylab="Log Odds", pch=16, cex=0.35, ...)

Arguments

fit 	 an MArrayLM fitted linear model object

coef 	 integer giving the coefficient

highlight 	 number of top genes to be highlighted

names 	 character vector giving text labels for the probes to be used in highlighting

xlab 	 character string giving label for x-axis

ylab 	 character string giving label for y-axis

pch 	 vector or list of plotting characters. Default is integer code 16 which gives a solid circle.

cex 	 numeric vector of plot symbol expansions. Default is 0.35.

... 	 any other arguments are passed to plot

Details

A volcano plot is any plot which displays fold changes versus a measure of statistical significance of the change.

Value

A plot is created on the current graphics device.

Author(s)

Gordon Smyth

See Also

An overview of presentation plots following the fitting of a linear model in LIMMA is given in 06.LinearModels.

Examples


#  See lmFit examples

[Package limma version 2.14.7 Index]