write.fit {limma}	R Documentation

Write MArrayLM Object to a File

Description

Write a microarray linear model fit to a file.

Usage

write.fit(fit, results = NULL, file, digits = NULL,
          adjust = "none", method = "separate", F.adjust = "none",
          quote = FALSE, sep = "\t", row.names = TRUE, ...)

Arguments

fit
object of class MArrayLM containing the results of a linear model fit.

results
object of class TestResults.

file
character string giving path name for the output file.

digits
integer indicating rounding precision for output values. If NULL, then no rounding is done.

adjust
character string specifying multiple-testing adjustment method for the t-statistic P-values, e.g., "BH". See p.adjust for the available options. If NULL or "none" then the P-values are not adjusted.

method
character string, should the P-value adjustment be "global" or "separate" for each contrast. Ignored if adjust = "none".

F.adjust
character string specifying adjustment method for the F-statistic P-values.

quote
logical value. If TRUE, any character or factor columns will be surrounded by double quotes.

sep
the field separator string. Values in the output file will be separated by this string.

row.names
logical value, whether to include row names in the output file.

...
other arguments are passed to write.table.

Details

This function writes a delimited text file containing for each gene (1) the average log2-intensity (AveExpr), (2) the coefficients or contrasts (log2-fold-changes, Coef), (3) moderated t-statistics, (4) t-statistic P-values, (5) F-statistic if available, (6) F-statistic P-values if available, (7) decideTests results if available and (8) gene names and annotation.

The results argument is optional. If supplied, it should be the output from decideTests for the same fit object, which indicates whether each contrast for each gene is considered statistically significant or not (coded 1 or -1 for positive or negative significant differences and 0 for non-significant values).

If fit contains row names and row.names=TRUE, then the row names will be the first column of the output file with a blank column heading. This behaviour is analogous to that of write.csv or to write.table with col.names=NA.

Value

No value is produced but a file is written to the current working directory.

Author(s)

Gordon Smyth

See Also

write.table or write.csv in the base library.

An overview of linear model functions in limma is given by 06.LinearModels.

Examples

## Not run: 
# The following three alternatives are equivalent:
write.fit(fit, file = "temp.csv", sep = ",")
write.csv(fit, file = "temp.csv")
a <- as.data.frame(fit)
write.csv(fit, file = "temp.csv")

## End(Not run)

[Package limma version 3.50.3 Index]