write.fit {limma}	R Documentation

Write MArrayLM Object to a File

Description

Write a microarray linear model fit to a file.

Usage


write.fit(fit, results=NULL, file, digits=3, adjust="none", method="separate", F.adjust="none", sep="\t", ...)

Arguments

fit 	 object of class MArrayLM containing the results of a linear model fit

results 	 object of class TestResults

file 	 character string giving name of file

digits 	 integer indicating precision to be used

adjust 	 character string specifying multiple-testing adjustment method for the t-statistic P-values, e.g., "BH". See p.adjust for the available options. If NULL or "none" then the P-values are not adjusted.

method 	 character string, should the P-value adjustment be "global" or "separate" for each contrast.

F.adjust 	 character string specifying adjustment method for the F-statistic P-values.

sep 	 the field separator string. Values in the output file will be separated by this string.

... 	 other arguments are passed to write.table

Details

This function writes a tab-delimited text file containing for each gene (1) the average log-intensity, (2) the log-ratios, (3) moderated t-statistics, (4) t-statistic P-values, (5) F-statistic if available, (6) F-statistic P-values if available, (7) classification if available and (8) gene names and annotation.

Value

No value is produced but a file is written to the current working directory.

Author(s)

Gordon Smyth

See Also

write in the base library.

An overview of linear model functions in limma is given by 06.LinearModels.

[Package limma version 2.14.7 Index]