zscore {limma}	R Documentation

Z-score Equivalents

Description

Compute z-score equivalents of for gamma or t-distribution random deviates.

Usage


zscoreGamma(q, shape, rate = 1, scale = 1/rate) 
zscoreT(x, df)
tZscore(x, df)

Arguments

q, x 	 numeric matrix for vector giving deviates of a random variaable

shape 	 gamma shape parameter (>0)

rate 	 gamma rate parameter (>0)

scale 	 gamma scale parameter (>0)

df 	 degrees of freedom (>0 for zscoreT or >=1 for tZscore)

Details

These functions compute the standard normal deviates which have the same quantiles as the given values in the specified distribution. For example, if z <- zscoreT(x,df=df) then pnorm(z) equals pt(x,df=df). tZscore is the inverse of zscoreT.

Care is taken to do the computations accurately in both tails of the distributions.

Value

Numeric vector giving equivalent deviates from a standard normal distribution (zscoreGamma and zscoreT) or deviates from a t-distribution (tZscore).

Author(s)

Gordon Smyth

See Also

qnorm, pgamma, pt

Examples


zscoreGamma(1, shape=1, scale=1)
zscoreT(2, df=3)
tZscore(2, df=3)

[Package limma version 2.14.7 Index]